refactor: decompose handlers.ts into per-domain handler modules
Split the 2879-line monolithic handlers.ts into focused per-domain modules
under a handlers/ directory:

- shared.ts: HandlerContext, types, utility functions (renderHistoryContent, etc.)
- sessions.ts: Session lifecycle, user messages, history, undo, regenerate
- skills.ts: Skill enable/disable/install/uninstall/update/search
- apps.ts: App CRUD, shared apps, gallery, bundling
- config.ts: Model, trust rules, schedules, reminders, integrations
- computer-use.ts: CU session create/abort/observation
- publish.ts: Vercel page publish/unpublish
- misc.ts: Task submit, suggestions, link open, blob probe
- index.ts: Dispatch map and re-exports

The original handlers.ts is now a thin re-export shim so all existing
imports continue to work without changes.
