+++ assistant/src/daemon/handlers.ts
@@ -1,3032 +1,17 @@
-import * as net from 'node:net';
-import { v4 as uuid } from 'uuid';
-import Anthropic from '@anthropic-ai/sdk';
-import { getConfig, loadRawConfig, saveRawConfig, invalidateConfigCache } from '../config/loader.js';
-import { getFailoverProvider, initializeProviders } from '../providers/registry.js';
-import { RateLimitProvider } from '../providers/ratelimit.js';
-import * as conversationStore from '../memory/conversation-store.js';
-import { getLogger } from '../util/logger.js';
-import { Session } from './session.js';
-import { ComputerUseSession } from './computer-use-session.js';
-import type {
-  ClientMessage,
-  ServerMessage,
-  ConfirmationResponse,
-  SecretResponse,
-  SessionCreateRequest,
-  SessionSwitchRequest,
-  CancelRequest,
-  ModelSetRequest,
-  HistoryRequest,
-  UndoRequest,
-  RegenerateRequest,
-  UsageRequest,
-  SandboxSetRequest,
-  UserMessage,
-  CuSessionCreate,
-  CuSessionAbort,
-  CuObservation,
-  TaskSubmit,
-  AppDataRequest,
-  SkillDetailRequest,
-  SkillsEnableRequest,
-  SkillsDisableRequest,
-  SkillsConfigureRequest,
-  SkillsInstallRequest,
-  SkillsUninstallRequest,
-  SkillsUpdateRequest,
-  SkillsCheckUpdatesRequest,
-  SkillsSearchRequest,
-  SkillsInspectRequest,
-  SuggestionRequest,
-  AddTrustRule,
-  RemoveTrustRule,
-  UpdateTrustRule,
-  ScheduleToggle,
-  ScheduleRemove,
-  ReminderCancel,
-  BundleAppRequest,
-  SharedAppDeleteRequest,
-  ForkSharedAppRequest,
-  ShareAppCloudRequest,
-  UiSurfaceShow,
-  IpcBlobProbe,
-  GalleryInstallRequest,
-  ShareToSlackRequest,
-  SlackWebhookConfigRequest,
-  VercelApiConfigRequest,
-  AppUpdatePreviewRequest,
-  PublishPageRequest,
-  UnpublishPageRequest,
-  LinkOpenRequest,
-} from './ipc-protocol.js';
-import { postToSlackWebhook } from '../slack/slack-webhook.js';
-import { createHash } from 'node:crypto';
-import { computeContentId } from '../util/content-id.js';
-import { execSync } from 'node:child_process';
-import { existsSync, rmSync, readdirSync, readFileSync } from 'node:fs';
-import { homedir } from 'node:os';
-import { join } from 'node:path';
-import { addRule, removeRule, updateRule, getAllRules } from '../permissions/trust-store.js';
-import { listSchedules, updateSchedule, deleteSchedule, describeCronExpression } from '../schedule/schedule-store.js';
-import { listReminders, cancelReminder } from '../tools/reminder/reminder-store.js';
-import { loadSkillCatalog, loadSkillBySelector, ensureSkillIcon } from '../config/skills.js';
-import { resolveSkillStates } from '../config/skill-state.js';
-import { handleRideShotgunStart } from './ride-shotgun-handler.js';
-import { handleWatchObservation } from './watch-handler.js';
-import { classifyInteraction } from './classifier.js';
-import { queryAppRecords, createAppRecord, updateAppRecord, deleteAppRecord, listApps, getApp, createApp, updateApp } from '../memory/app-store.js';
-import { defaultGallery } from '../gallery/default-gallery.js';
-import { getWorkspaceSkillsDir } from '../util/platform.js';
-import { clawhubInstall, clawhubUpdate, clawhubSearch, clawhubCheckUpdates, clawhubInspect } from '../skills/clawhub.js';
-import { parseSlashCandidate } from '../skills/slash-commands.js';
-import { removeSkillsIndexEntry, deleteManagedSkill, validateManagedSkillId } from '../skills/managed-store.js';
-import { packageApp } from '../bundler/app-bundler.js';
-import { createSharedAppLink } from '../memory/shared-app-links-store.js';
-import { handleOpenBundle } from './handlers/open-bundle-handler.js';
-import { checkIngressForSecrets } from '../security/secret-ingress.js';
-import { deployHtmlToVercel, deleteVercelDeployment } from '../services/vercel-deploy.js';
-import { createPublishedPage, getPublishedPageByHash, markDeleted, getPublishedPageByDeploymentId } from '../memory/published-pages-store.js';
-import { getSecureKey, setSecureKey, deleteSecureKey } from '../security/secure-keys.js';
-import { upsertCredentialMetadata, deleteCredentialMetadata } from '../tools/credentials/metadata-store.js';
-import { credentialBroker } from '../tools/credentials/broker.js';
-import type { SecretPromptResult } from '../permissions/secret-prompter.js';
-import { resolveBlobPath, readBlob, deleteBlob, isValidBlobId, validateBlobKindEncoding } from './ipc-blob-store.js';
-import { estimateBase64Bytes } from './assistant-attachments.js';
-import { classifySessionError, buildSessionErrorMessage } from './session-error.js';
-import { getAttachmentsForMessageUnscoped } from '../memory/attachments-store.js';
-import type { UserMessageAttachment } from './ipc-contract.js';
-import { listStatuses, getIntegration, disconnect as disconnectIntegration } from '../integrations/registry.js';
-import { startOAuth2Flow } from '../integrations/oauth2.js';
-import type { IntegrationConnectRequest, IntegrationDisconnectRequest } from './ipc-contract.js';
-
-const log = getLogger('handlers');
-const HISTORY_ATTACHMENT_TEXT_LIMIT = 500;
-
-// Module-level map for non-session secret prompts (e.g. publish_page)
-const pendingStandaloneSecrets = new Map<string, { resolve: (result: SecretPromptResult) => void; timer: ReturnType<typeof setTimeout> }>();
-
-const FALLBACK_SCREEN = { width: 1920, height: 1080 };
-let cachedScreenDims: { width: number; height: number } | null = null;
-const cuObservationSequenceBySession = new Map<string, number>();
-
-/**
- * Query the main display dimensions via CoreGraphics.
- * Cached after the first successful call; falls back to 1920x1080.
- */
-function getScreenDimensions(): { width: number; height: number } {
-  if (cachedScreenDims) return cachedScreenDims;
-  try {
-    const out = execSync(
-      `swift -e 'import CoreGraphics; let b = CGDisplayBounds(CGMainDisplayID()); print("\\(Int(b.width))x\\(Int(b.height))")'`,
-      { timeout: 10_000, encoding: 'utf-8' },
-    ).trim();
-    const [w, h] = out.split('x').map(Number);
-    if (w > 0 && h > 0) {
-      cachedScreenDims = { width: w, height: h };
-      return cachedScreenDims;
-    }
-  } catch (err) {
-    log.debug({ err }, 'Failed to query screen dimensions, using fallback');
-  }
-  return FALLBACK_SCREEN;
-}
-
-/**
- * Find the current socket bound to a given session by reversing the
- * `socketToSession` map. Returns `undefined` if no socket is bound.
- */
-function findSocketForSession(
-  sessionId: string,
-  ctx: HandlerContext,
-): net.Socket | undefined {
-  for (const [sock, id] of ctx.socketToSession) {
-    if (id === sessionId) return sock;
-  }
-  return undefined;
-}
-
-/**
- * Wire the escalation handler on a text_qa session so that invoking
- * `request_computer_control` creates a CU session and notifies the client.
- *
- * Instead of closing over the original `socket`, the handler looks up the
- * current socket for the session at call time via `ctx.socketToSession`.
- * This ensures the handler targets the correct socket even after a
- * disconnect-and-rebind cycle.
- */
-function wireEscalationHandler(
-  session: Session,
-  _socket: net.Socket,
-  ctx: HandlerContext,
-  explicitWidth?: number,
-  explicitHeight?: number,
-): void {
-  const dims = (explicitWidth && explicitHeight)
-    ? { width: explicitWidth, height: explicitHeight }
-    : getScreenDimensions();
-  const screenWidth = dims.width;
-  const screenHeight = dims.height;
-  session.setEscalationHandler((task: string, sourceSessionId: string): boolean => {
-    const currentSocket = findSocketForSession(sourceSessionId, ctx);
-    if (!currentSocket) {
-      log.warn({ sourceSessionId }, 'Escalation handler: no active socket found for session');
-      return false;
-    }
-
-    const cuSessionId = uuid();
-    const cuMsg: CuSessionCreate = {
-      type: 'cu_session_create',
-      sessionId: cuSessionId,
-      task,
-      screenWidth,
-      screenHeight,
-      interactionType: 'computer_use',
-    };
-    handleCuSessionCreate(cuMsg, currentSocket, ctx);
-
-    ctx.send(currentSocket, {
-      type: 'task_routed',
-      sessionId: cuSessionId,
-      interactionType: 'computer_use',
-      task,
-      escalatedFrom: sourceSessionId,
-    });
-
-    return true;
-  });
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === 'object' && value !== null;
-}
-
-function formatBytes(sizeBytes: number): string {
-  if (sizeBytes < 1024) return `${sizeBytes} B`;
-  const kb = sizeBytes / 1024;
-  if (kb < 1024) return `${kb.toFixed(1)} KB`;
-  return `${(kb / 1024).toFixed(1)} MB`;
-}
-
-function clampAttachmentText(text: string): string {
-  if (text.length <= HISTORY_ATTACHMENT_TEXT_LIMIT) return text;
-  return `${text.slice(0, HISTORY_ATTACHMENT_TEXT_LIMIT)}<truncated />`;
-}
-
-function renderImageBlockForHistory(block: Record<string, unknown>): string {
-  const source = isRecord(block.source) ? block.source : null;
-  const mediaType = source && typeof source.media_type === 'string' ? source.media_type : 'image/*';
-  const sizeBytes = source && typeof source.data === 'string' ? estimateBase64Bytes(source.data) : 0;
-  if (sizeBytes <= 0) {
-    return `[Image attachment] ${mediaType}`;
-  }
-  return `[Image attachment] ${mediaType}, ${formatBytes(sizeBytes)}`;
-}
-
-function renderFileBlockForHistory(block: Record<string, unknown>): string {
-  const source = isRecord(block.source) ? block.source : null;
-  const mediaType = source && typeof source.media_type === 'string' ? source.media_type : 'application/octet-stream';
-  const filename = source && typeof source.filename === 'string' ? source.filename : 'attachment';
-  const sizeBytes = source && typeof source.data === 'string' ? estimateBase64Bytes(source.data) : 0;
-  const summaryParts = [`[File attachment] ${filename}`, `type=${mediaType}`];
-  if (sizeBytes > 0) summaryParts.push(`size=${formatBytes(sizeBytes)}`);
-
-  const extractedText = typeof block.extracted_text === 'string' ? block.extracted_text.trim() : '';
-  if (!extractedText) {
-    return summaryParts.join(', ');
-  }
-  return `${summaryParts.join(', ')}\nAttachment text: ${clampAttachmentText(extractedText)}`;
-}
-
-export interface HistoryToolCall {
-  name: string;
-  input: Record<string, unknown>;
-  result?: string;
-  isError?: boolean;
-  /** Base64-encoded image data from tool contentBlocks (e.g. browser_screenshot). */
-  imageData?: string;
-}
-
-export interface HistorySurface {
-  surfaceId: string;
-  surfaceType: string;
-  title?: string;
-  data: Record<string, unknown>;
-  actions?: Array<{ id: string; label: string; style?: string }>;
-  display?: string;
-}
-
-export interface RenderedHistoryContent {
-  text: string;
-  toolCalls: HistoryToolCall[];
-  /** True when the first tool_use block appeared before any text block. */
-  toolCallsBeforeText: boolean;
-  /** Text segments split by tool-call boundaries. */
-  textSegments: string[];
-  /** Content block ordering using "text:N", "tool:N", "surface:N" encoding. */
-  contentOrder: string[];
-  /** UI surfaces (widgets) embedded in the message. */
-  surfaces: HistorySurface[];
-}
-
-export function renderHistoryContent(content: unknown): RenderedHistoryContent {
-  if (!Array.isArray(content)) {
-    let text: string;
-    if (content == null) {
-      text = '';
-    } else if (typeof content === 'object') {
-      text = JSON.stringify(content);
-    } else {
-      text = String(content);
-    }
-    return {
-      text,
-      toolCalls: [],
-      toolCallsBeforeText: false,
-      textSegments: text ? [text] : [],
-      contentOrder: text ? ['text:0'] : [],
-      surfaces: [],
-    };
-  }
-
-  const textParts: string[] = [];
-  const attachmentParts: string[] = [];
-  const toolCalls: HistoryToolCall[] = [];
-  const surfaces: HistorySurface[] = [];
-  const pendingToolUses = new Map<string, HistoryToolCall>();
-  let seenText = false;
-  let seenToolUse = false;
-  let toolCallsBeforeText = false;
-
-  // Segment tracking: text blocks separated by tool_use boundaries
-  const textSegments: string[] = [];
-  const contentOrder: string[] = [];
-  let currentSegmentParts: string[] = [];
-  let hasOpenSegment = false;
-
-  function finalizeSegment(): void {
-    if (hasOpenSegment) {
-      textSegments[textSegments.length - 1] = currentSegmentParts.join('');
-      currentSegmentParts = [];
-      hasOpenSegment = false;
-    }
-  }
-
-  function ensureSegment(): void {
-    if (!hasOpenSegment) {
-      textSegments.push('');
-      contentOrder.push(`text:${textSegments.length - 1}`);
-      hasOpenSegment = true;
-    }
-  }
-
-  for (const block of content) {
-    if (!isRecord(block) || typeof block.type !== 'string') continue;
-
-    // Collect ui_surface blocks for inclusion in history
-    if (block.type === 'ui_surface') {
-      finalizeSegment();
-      const surface: HistorySurface = {
-        surfaceId: typeof block.surfaceId === 'string' ? block.surfaceId : '',
-        surfaceType: typeof block.surfaceType === 'string' ? block.surfaceType : '',
-        title: typeof block.title === 'string' ? block.title : undefined,
-        data: isRecord(block.data) ? (block.data as Record<string, unknown>) : {},
-        actions: Array.isArray(block.actions) ? block.actions : undefined,
-        display: typeof block.display === 'string' ? block.display : undefined,
-      };
-      surfaces.push(surface);
-      contentOrder.push(`surface:${surfaces.length - 1}`);
-      continue;
-    }
-
-    if (block.type === 'text' && typeof block.text === 'string') {
-      textParts.push(block.text);
-      ensureSegment();
-      currentSegmentParts.push(block.text);
-      seenText = true;
-      continue;
-    }
-    if (block.type === 'file') {
-      attachmentParts.push(renderFileBlockForHistory(block));
-      continue;
-    }
-    if (block.type === 'image') {
-      attachmentParts.push(renderImageBlockForHistory(block));
-      continue;
-    }
-    if (block.type === 'tool_use') {
-      finalizeSegment();
-      const name = typeof block.name === 'string' ? block.name : 'unknown';
-      const input = isRecord(block.input) ? block.input as Record<string, unknown> : {};
-      const id = typeof block.id === 'string' ? block.id : '';
-      const entry: HistoryToolCall = { name, input };
-      toolCalls.push(entry);
-      if (id) pendingToolUses.set(id, entry);
-      contentOrder.push(`tool:${toolCalls.length - 1}`);
-      if (!seenToolUse) {
-        seenToolUse = true;
-        if (!seenText) toolCallsBeforeText = true;
-      }
-      continue;
-    }
-    if (block.type === 'tool_result') {
-      const toolUseId = typeof block.tool_use_id === 'string' ? block.tool_use_id : '';
-      const resultContent = typeof block.content === 'string' ? block.content : '';
-      const isError = block.is_error === true;
-      // Extract base64 image data from persisted contentBlocks (e.g. browser_screenshot)
-      let imageData: string | undefined;
-      if (Array.isArray(block.contentBlocks)) {
-        const imgBlock = block.contentBlocks.find(
-          (b: Record<string, unknown>) => isRecord(b) && b.type === 'image',
-        );
-        if (imgBlock && isRecord(imgBlock) && isRecord(imgBlock.source)) {
-          const src = imgBlock.source as Record<string, unknown>;
-          if (typeof src.data === 'string') {
-            imageData = src.data;
-          }
-        }
-      }
-      const matched = toolUseId ? pendingToolUses.get(toolUseId) : null;
-      if (matched) {
-        matched.result = resultContent;
-        matched.isError = isError;
-        if (imageData) matched.imageData = imageData;
-      } else {
-        toolCalls.push({ name: 'unknown', input: {}, result: resultContent, isError, ...(imageData ? { imageData } : {}) });
-      }
-      continue;
-    }
-  }
-
-  // Attachment descriptions are NOT included in textSegments — they are
-  // rendered separately as attachment chips/images in the UI.  Including
-  // them would cause the "last text segment" heuristic to surface an
-  // attachment descriptor instead of the assistant's actual narrative text
-  // in interleaved history messages.
-
-  finalizeSegment();
-
-  const text = textParts.join('');
-  let rendered: string;
-  if (attachmentParts.length === 0) {
-    rendered = text;
-  } else if (text.trim().length === 0) {
-    rendered = attachmentParts.join('\n');
-  } else {
-    rendered = `${text}\n${attachmentParts.join('\n')}`;
-  }
-
-  return { text: rendered, toolCalls, toolCallsBeforeText, textSegments, contentOrder, surfaces };
-}
-
-/**
- * Optional overrides for session creation (e.g. interview mode).
- */
-export interface SessionCreateOptions {
-  systemPromptOverride?: string;
-  maxResponseTokens?: number;
-}
-
-/**
- * Shared context that handlers need from the DaemonServer.
- * Keeps handlers decoupled from the server class itself.
- */
-export interface HandlerContext {
-  sessions: Map<string, Session>;
-  socketToSession: Map<net.Socket, string>;
-  cuSessions: Map<string, ComputerUseSession>;
-  socketToCuSession: Map<net.Socket, Set<string>>;
-  cuObservationParseSequence: Map<string, number>;
-  socketSandboxOverride: Map<net.Socket, boolean>;
-  sharedRequestTimestamps: number[];
-  debounceTimers: Map<string, ReturnType<typeof setTimeout>>;
-  suppressConfigReload: boolean;
-  setSuppressConfigReload(value: boolean): void;
-  updateConfigFingerprint(): void;
-  send(socket: net.Socket, msg: ServerMessage): void;
-  broadcast(msg: ServerMessage): void;
-  clearAllSessions(): number;
-  getOrCreateSession(
-    conversationId: string,
-    socket?: net.Socket,
-    rebindClient?: boolean,
-    options?: SessionCreateOptions,
-  ): Promise<Session>;
-}
-
-// ─── Typed dispatch ──────────────────────────────────────────────────────────
-
-type MessageType = ClientMessage['type'];
-type MessageOfType<T extends MessageType> = Extract<ClientMessage, { type: T }>;
-type MessageHandler<T extends MessageType> = (
-  msg: MessageOfType<T>,
-  socket: net.Socket,
-  ctx: HandlerContext,
-) => void | Promise<void>;
-type DispatchMap = { [T in MessageType]: MessageHandler<T> };
-
-const handlers: DispatchMap = {
-  user_message: handleUserMessage,
-  confirmation_response: handleConfirmationResponse,
-  secret_response: handleSecretResponse,
-  session_list: (_msg, socket, ctx) => handleSessionList(socket, ctx),
-  session_create: handleSessionCreate,
-  sessions_clear: (_msg, socket, ctx) => handleSessionsClear(socket, ctx),
-  session_switch: handleSessionSwitch,
-  cancel: handleCancel,
-  model_get: (_msg, socket, ctx) => handleModelGet(socket, ctx),
-  model_set: handleModelSet,
-  history_request: handleHistoryRequest,
-  undo: handleUndo,
-  regenerate: handleRegenerate,
-  usage_request: handleUsageRequest,
-  sandbox_set: handleSandboxSet,
-  cu_session_create: handleCuSessionCreate,
-  cu_session_abort: handleCuSessionAbort,
-  cu_observation: handleCuObservation,
-  ride_shotgun_start: handleRideShotgunStart,
-  watch_observation: handleWatchObservation,
-  task_submit: handleTaskSubmit,
-  app_data_request: handleAppDataRequest,
-  skills_list: (_msg, socket, ctx) => handleSkillsList(socket, ctx),
-  skill_detail: handleSkillDetail,
-  skills_enable: handleSkillsEnable,
-  skills_disable: handleSkillsDisable,
-  skills_configure: handleSkillsConfigure,
-  skills_install: handleSkillsInstall,
-  skills_uninstall: handleSkillsUninstall,
-  skills_update: handleSkillsUpdate,
-  skills_check_updates: handleSkillsCheckUpdates,
-  skills_search: handleSkillsSearch,
-  skills_inspect: handleSkillsInspect,
-  suggestion_request: handleSuggestionRequest,
-  add_trust_rule: handleAddTrustRule,
-  trust_rules_list: (_msg, socket, ctx) => handleTrustRulesList(socket, ctx),
-  remove_trust_rule: handleRemoveTrustRule,
-  update_trust_rule: handleUpdateTrustRule,
-  schedules_list: (_msg, socket, ctx) => handleSchedulesList(socket, ctx),
-  schedule_toggle: handleScheduleToggle,
-  schedule_remove: handleScheduleRemove,
-  reminders_list: (_msg, socket, ctx) => handleRemindersList(socket, ctx),
-  reminder_cancel: handleReminderCancel,
-  share_app_cloud: handleShareAppCloud,
-  bundle_app: handleBundleApp,
-  open_bundle: handleOpenBundle,
-  app_open_request: (msg, socket, ctx) => handleAppOpenRequest(msg, socket, ctx),
-  app_update_preview: handleAppUpdatePreview,
-  apps_list: (_msg, socket, ctx) => handleAppsList(socket, ctx),
-  shared_apps_list: (_msg, socket, ctx) => handleSharedAppsList(socket, ctx),
-  shared_app_delete: handleSharedAppDelete,
-  fork_shared_app: handleForkSharedApp,
-  sign_bundle_payload_response: (_msg, _socket, _ctx) => {
-    // TODO(signing): Route to pending promise resolution once the daemon-driven
-    // IPC signing orchestration is wired up. Currently a no-op placeholder to
-    // satisfy the exhaustive dispatch map; signing is invoked via SigningCallback.
-  },
-  get_signing_identity_response: (_msg, _socket, _ctx) => {
-    // TODO(signing): Route to pending promise resolution once the daemon-driven
-    // IPC signing orchestration is wired up. Currently a no-op placeholder to
-    // satisfy the exhaustive dispatch map; signing is invoked via SigningCallback.
-  },
-  gallery_list: (_msg, socket, ctx) => handleGalleryList(socket, ctx),
-  gallery_install: handleGalleryInstall,
-  share_to_slack: handleShareToSlack,
-  slack_webhook_config: handleSlackWebhookConfig,
-  vercel_api_config: handleVercelApiConfig,
-  publish_page: handlePublishPage,
-  unpublish_page: handleUnpublishPage,
-  ping: (_msg, socket, ctx) => { ctx.send(socket, { type: 'pong' }); },
-  link_open_request: handleLinkOpenRequest,
-  ipc_blob_probe: handleIpcBlobProbe,
-  ui_surface_action: (msg, _socket, ctx) => {
-    const cuSession = ctx.cuSessions.get(msg.sessionId);
-    if (cuSession) {
-      cuSession.handleSurfaceAction(msg.surfaceId, msg.actionId, msg.data);
-      return;
-    }
-    const session = ctx.sessions.get(msg.sessionId);
-    if (session) {
-      session.handleSurfaceAction(msg.surfaceId, msg.actionId, msg.data);
-      return;
-    }
-    log.warn({ sessionId: msg.sessionId, surfaceId: msg.surfaceId }, 'No session found for surface action');
-  },
-  ui_surface_undo: (msg, _socket, ctx) => {
-    const session = ctx.sessions.get(msg.sessionId);
-    if (session) {
-      session.handleSurfaceUndo(msg.surfaceId);
-      return;
-    }
-    log.warn({ sessionId: msg.sessionId, surfaceId: msg.surfaceId }, 'No session found for surface undo');
-  },
-  integration_list: (_msg, socket, ctx) => {
-    handleIntegrationList(socket, ctx);
-  },
-  integration_connect: (msg, socket, ctx) => {
-    handleIntegrationConnect(msg as IntegrationConnectRequest, socket, ctx);
-  },
-  integration_disconnect: (msg, socket, ctx) => {
-    handleIntegrationDisconnect(msg as IntegrationDisconnectRequest, socket, ctx);
-  },
-};
-
-export function handleMessage(
-  msg: ClientMessage,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  const handler = handlers[msg.type] as
-    | ((msg: ClientMessage, socket: net.Socket, ctx: HandlerContext) => void)
-    | undefined;
-  if (!handler) {
-    log.warn({ type: msg.type }, 'Unknown message type, ignoring');
-    return;
-  }
-  handler(msg, socket, ctx);
-}
-
-// ─── Individual handlers ─────────────────────────────────────────────────────
-
-async function handleUserMessage(
-  msg: UserMessage,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): Promise<void> {
-  const requestId = uuid();
-  const rlog = log.child({ sessionId: msg.sessionId, requestId });
-  try {
-    ctx.socketToSession.set(socket, msg.sessionId);
-    const session = await ctx.getOrCreateSession(msg.sessionId, socket, true);
-    // Only wire the escalation handler if one isn't already set — handleTaskSubmit
-    // sets a handler with the client's actual screen dimensions, and overwriting it
-    // here would replace those dimensions with the daemon's defaults.
-    if (!session.hasEscalationHandler()) {
-      wireEscalationHandler(session, socket, ctx);
-    }
-
-    const sendEvent = (event: ServerMessage) => ctx.send(socket, event);
-
-    // Block inbound messages that contain secrets and redirect to secure prompt
-    const ingressCheck = checkIngressForSecrets(msg.content ?? '');
-    if (ingressCheck.blocked) {
-      rlog.warn({ detectedTypes: ingressCheck.detectedTypes }, 'Blocked user message containing secrets');
-      ctx.send(socket, {
-        type: 'error',
-        message: ingressCheck.userNotice!,
-      });
-      // Redirect: trigger a secure prompt so the user can enter the secret safely
-      session.redirectToSecurePrompt(ingressCheck.detectedTypes);
-      return;
-    }
-
-    session.traceEmitter.emit('request_received', 'User message received', {
-      requestId,
-      status: 'info',
-      attributes: { source: 'user_message' },
-    });
-
-    const result = session.enqueueMessage(msg.content ?? '', msg.attachments ?? [], sendEvent, requestId, msg.activeSurfaceId, msg.currentPage);
-    if (result.rejected) {
-      rlog.warn('Message rejected — queue is full');
-      session.traceEmitter.emit('request_error', 'Message rejected — queue is full', {
-        requestId,
-        status: 'error',
-        attributes: { reason: 'queue_full', queueDepth: session.getQueueDepth() },
-      });
-      ctx.send(socket, buildSessionErrorMessage(msg.sessionId, {
-        code: 'QUEUE_FULL',
-        userMessage: 'Message queue is full (max depth: 10). Please wait for current messages to be processed.',
-        retryable: true,
-        debugDetails: 'Message rejected — session queue is full',
-      }));
-      return;
-    }
-    if (result.queued) {
-      const position = session.getQueueDepth();
-      rlog.info({ position }, 'Message queued (session busy)');
-      session.traceEmitter.emit('request_queued', `Message queued at position ${position}`, {
-        requestId,
-        status: 'info',
-        attributes: { position },
-      });
-      ctx.send(socket, {
-        type: 'message_queued',
-        sessionId: msg.sessionId,
-        requestId,
-        position,
-      });
-      return; // Don't await — message will be processed when current one finishes
-    }
-
-    rlog.info('Processing user message');
-    // Fire-and-forget: don't block the IPC handler so the connection can
-    // continue receiving messages (e.g. cancel, confirmations, or
-    // additional user_message that will be queued by the session).
-    session.processMessage(msg.content ?? '', msg.attachments ?? [], sendEvent, requestId, msg.activeSurfaceId, msg.currentPage).catch((err) => {
-      const message = err instanceof Error ? err.message : String(err);
-      rlog.error({ err }, 'Error processing user message (session or provider failure)');
-      ctx.send(socket, { type: 'error', message: `Failed to process message: ${message}` });
-      const classified = classifySessionError(err, { phase: 'agent_loop' });
-      ctx.send(socket, buildSessionErrorMessage(msg.sessionId, classified));
-    });
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    rlog.error({ err }, 'Error setting up user message processing');
-    ctx.send(socket, { type: 'error', message: `Failed to process message: ${message}` });
-    const classified = classifySessionError(err, { phase: 'handler' });
-    ctx.send(socket, buildSessionErrorMessage(msg.sessionId, classified));
-  }
-}
-
-function handleConfirmationResponse(
-  msg: ConfirmationResponse,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  const sessionId = ctx.socketToSession.get(socket);
-  if (sessionId) {
-    const session = ctx.sessions.get(sessionId);
-    if (session) {
-      session.handleConfirmationResponse(
-        msg.requestId,
-        msg.decision as 'allow' | 'always_allow' | 'deny',
-        msg.selectedPattern,
-        msg.selectedScope,
-      );
-    }
-  }
-}
-
-function handleSecretResponse(
-  msg: SecretResponse,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  // Check standalone (non-session) prompts first, since they use a dedicated
-  // requestId that won't collide with session prompts.
-  const standalone = pendingStandaloneSecrets.get(msg.requestId);
-  if (standalone) {
-    clearTimeout(standalone.timer);
-    pendingStandaloneSecrets.delete(msg.requestId);
-    standalone.resolve({ value: msg.value ?? null, delivery: msg.delivery ?? 'store' });
-    return;
-  }
-
-  // Otherwise route to the session's secret prompter
-  const sessionId = ctx.socketToSession.get(socket);
-  if (sessionId) {
-    const session = ctx.sessions.get(sessionId);
-    if (session) {
-      session.handleSecretResponse(msg.requestId, msg.value, msg.delivery);
-    }
-  }
-}
-
-/**
- * Send a `secret_request` to the client and wait for the response,
- * outside of a session context (e.g. from handler-level code like publish_page).
- */
-function requestSecretStandalone(
-  socket: net.Socket,
-  ctx: HandlerContext,
-  params: { service: string; field: string; label: string; description?: string; placeholder?: string; allowedTools?: string[] },
-): Promise<SecretPromptResult> {
-  const requestId = uuid();
-  return new Promise((resolve) => {
-    const timer = setTimeout(() => {
-      pendingStandaloneSecrets.delete(requestId);
-      resolve({ value: null, delivery: 'store' });
-    }, 120_000);
-    pendingStandaloneSecrets.set(requestId, { resolve, timer });
-    ctx.send(socket, {
-      type: 'secret_request',
-      requestId,
-      service: params.service,
-      field: params.field,
-      label: params.label,
-      description: params.description,
-      placeholder: params.placeholder,
-      purpose: 'Publish site apps to the web',
-      allowedTools: params.allowedTools,
-    });
-  });
-}
-
-function handleSessionList(socket: net.Socket, ctx: HandlerContext): void {
-  const conversations = conversationStore.listConversations(50);
-  ctx.send(socket, {
-    type: 'session_list_response',
-    sessions: conversations.map((c) => ({
-      id: c.id,
-      title: c.title ?? 'Untitled',
-      updatedAt: c.updatedAt,
-    })),
-  });
-}
-
-function handleSessionsClear(socket: net.Socket, ctx: HandlerContext): void {
-  const cleared = ctx.clearAllSessions();
-  // Also clear DB conversations. When a new IPC connection triggers
-  // sendInitialSession, it auto-creates a conversation if none exist.
-  // Without this DB clear, that auto-created row survives, contradicting
-  // the "clear all conversations" intent.
-  conversationStore.clearAll();
-  ctx.send(socket, { type: 'sessions_clear_response', cleared });
-}
-
-async function handleSessionCreate(
-  msg: SessionCreateRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): Promise<void> {
-  const conversation = conversationStore.createConversation(
-    msg.title ?? 'New Conversation',
-  );
-  const session = await ctx.getOrCreateSession(conversation.id, socket, true, {
-    systemPromptOverride: msg.systemPromptOverride,
-    maxResponseTokens: msg.maxResponseTokens,
-  });
-  wireEscalationHandler(session, socket, ctx);
-  ctx.send(socket, {
-    type: 'session_info',
-    sessionId: conversation.id,
-    title: conversation.title ?? 'New Conversation',
-    ...(msg.correlationId ? { correlationId: msg.correlationId } : {}),
-  });
-}
-
-async function handleSessionSwitch(
-  msg: SessionSwitchRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): Promise<void> {
-  const conversation = conversationStore.getConversation(msg.sessionId);
-  if (!conversation) {
-    ctx.send(socket, { type: 'error', message: `Session ${msg.sessionId} not found` });
-    return;
-  }
-  ctx.socketToSession.set(socket, msg.sessionId);
-  const session = await ctx.getOrCreateSession(msg.sessionId, socket, true);
-  // Only wire the escalation handler if one isn't already set — handleTaskSubmit
-  // sets a handler with the client's actual screen dimensions, and overwriting it
-  // here would replace those dimensions with the daemon's defaults.
-  if (!session.hasEscalationHandler()) {
-    wireEscalationHandler(session, socket, ctx);
-  }
-  ctx.send(socket, {
-    type: 'session_info',
-    sessionId: conversation.id,
-    title: conversation.title ?? 'Untitled',
-  });
-}
-
-function handleCancel(msg: CancelRequest, socket: net.Socket, ctx: HandlerContext): void {
-  const sessionId = msg.sessionId || ctx.socketToSession.get(socket);
-  if (sessionId) {
-    const session = ctx.sessions.get(sessionId);
-    if (session) {
-      session.abort();
-    }
-  }
-}
-
-function handleModelGet(socket: net.Socket, ctx: HandlerContext): void {
-  const config = getConfig();
-  ctx.send(socket, {
-    type: 'model_info',
-    model: config.model,
-    provider: config.provider,
-  });
-}
-
-function handleModelSet(
-  msg: ModelSetRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  try {
-    // Use raw config to avoid persisting env-var API keys to disk
-    const raw = loadRawConfig();
-    raw.model = msg.model;
-
-    // Suppress the file watcher callback — handleModelSet already does
-    // the full reload sequence; a redundant watcher-triggered reload
-    // would incorrectly evict sessions created after this method returns.
-    ctx.setSuppressConfigReload(true);
-    try {
-      saveRawConfig(raw);
-    } catch (err) {
-      ctx.setSuppressConfigReload(false);
-      throw err;
-    }
-    const existingSuppressTimer = ctx.debounceTimers.get('__suppress_reset__');
-    if (existingSuppressTimer) clearTimeout(existingSuppressTimer);
-    const resetTimer = setTimeout(() => { ctx.setSuppressConfigReload(false); }, 300);
-    ctx.debounceTimers.set('__suppress_reset__', resetTimer);
-
-    // Re-initialize provider with the new model so LLM calls use it
-    const config = getConfig();
-    initializeProviders(config);
-
-    // Evict idle sessions immediately; mark busy ones as stale so they
-    // get recreated with the new provider once they finish processing.
-    for (const [id, session] of ctx.sessions) {
-      if (!session.isProcessing()) {
-        session.dispose();
-        ctx.sessions.delete(id);
-      } else {
-        session.markStale();
-      }
-    }
-
-    ctx.updateConfigFingerprint();
-
-    ctx.send(socket, {
-      type: 'model_info',
-      model: config.model,
-      provider: config.provider,
-    });
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    ctx.send(socket, { type: 'error', message: `Failed to set model: ${message}` });
-  }
-}
-
-function handleSandboxSet(
-  msg: SandboxSetRequest,
-  _socket: net.Socket,
-  _ctx: HandlerContext,
-): void {
-  log.warn(
-    { enabled: msg.enabled },
-    'Received deprecated sandbox_set message. Runtime sandbox overrides are ignored.',
-  );
-}
-
-export interface ParsedHistoryMessage {
-  id?: string;
-  role: string;
-  text: string;
-  timestamp: number;
-  toolCalls: HistoryToolCall[];
-  toolCallsBeforeText: boolean;
-  textSegments: string[];
-  contentOrder: string[];
-  surfaces: HistorySurface[];
-}
-
-function handleHistoryRequest(
-  msg: HistoryRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  const dbMessages = conversationStore.getMessages(msg.sessionId);
-  const parsed: ParsedHistoryMessage[] = dbMessages.map((m) => {
-    let text = '';
-    let toolCalls: HistoryToolCall[] = [];
-    let toolCallsBeforeText = false;
-    let textSegments: string[] = [];
-    let contentOrder: string[] = [];
-    let surfaces: HistorySurface[] = [];
-    try {
-      const content = JSON.parse(m.content);
-      const rendered = renderHistoryContent(content);
-      text = rendered.text;
-      toolCalls = rendered.toolCalls;
-      toolCallsBeforeText = rendered.toolCallsBeforeText;
-      textSegments = rendered.textSegments;
-      contentOrder = rendered.contentOrder;
-      surfaces = rendered.surfaces;
-      if (m.role === 'assistant' && toolCalls.length > 0) {
-        log.info({ messageId: m.id, toolCallCount: toolCalls.length, text: text.substring(0, 100) }, 'History message with tool calls');
-      }
-    } catch (err) {
-      log.debug({ err, messageId: m.id }, 'Failed to parse message content as JSON, using raw text');
-      text = m.content;
-      textSegments = text ? [text] : [];
-      contentOrder = text ? ['text:0'] : [];
-      surfaces = [];
-    }
-    return { id: m.id, role: m.role, text, timestamp: m.createdAt, toolCalls, toolCallsBeforeText, textSegments, contentOrder, surfaces };
-  });
-
-  // Merge tool_result data from user messages into the preceding assistant
-  // message's toolCalls, and suppress user messages that only contain
-  // tool_result blocks (internal agent-loop turns).
-  const merged = mergeToolResults(parsed);
-
-  const historyMessages = merged.map((m) => {
-    let attachments: UserMessageAttachment[] | undefined;
-    if (m.role === 'assistant' && m.id) {
-      const linked = getAttachmentsForMessageUnscoped(m.id);
-      if (linked.length > 0) {
-        attachments = linked.map((a) => ({
-          id: a.id,
-          filename: a.originalFilename,
-          mimeType: a.mimeType,
-          data: a.dataBase64,
-        }));
-      }
-    }
-    return {
-      ...(m.id ? { id: m.id } : {}),
-      role: m.role,
-      text: m.text,
-      timestamp: m.timestamp,
-      ...(m.toolCalls.length > 0 ? { toolCalls: m.toolCalls, toolCallsBeforeText: m.toolCallsBeforeText } : {}),
-      ...(attachments ? { attachments } : {}),
-      ...(m.textSegments.length > 0 ? { textSegments: m.textSegments } : {}),
-      ...(m.contentOrder.length > 0 ? { contentOrder: m.contentOrder } : {}),
-      ...(m.surfaces.length > 0 ? { surfaces: m.surfaces } : {}),
-    };
-  });
-  ctx.send(socket, { type: 'history_response', sessionId: msg.sessionId, messages: historyMessages });
-
-  // Surfaces are now included directly in the history_response message (in the surfaces array),
-  // so we no longer emit separate ui_surface_show messages during history loading.
-}
-
-export function mergeToolResults(messages: ParsedHistoryMessage[]): ParsedHistoryMessage[] {
-  // Note: We no longer merge consecutive assistant messages at load time since
-  // they are now consolidated when saved. This function only handles legacy
-  // conversations that haven't been consolidated yet, and continues to merge
-  // tool_result blocks into their preceding assistant messages.
-  const result: ParsedHistoryMessage[] = [];
-  for (let i = 0; i < messages.length; i++) {
-    const msg = messages[i];
-
-    // If this is a user message whose only content is tool_result blocks,
-    // merge those results into the preceding assistant message's toolCalls.
-    // This should rarely happen now since consolidation removes these messages,
-    // but we keep it for backwards compatibility with old data.
-    if (msg.role === 'user' && msg.text.trim() === '' && msg.toolCalls.length > 0) {
-      const prev = result.length > 0 ? result[result.length - 1] : null;
-      if (prev && prev.role === 'assistant' && prev.toolCalls.length > 0) {
-        for (const resultEntry of msg.toolCalls) {
-          const unresolved = prev.toolCalls.find(
-            (tc) => tc.result === undefined,
-          );
-          if (unresolved) {
-            unresolved.result = resultEntry.result;
-            unresolved.isError = resultEntry.isError;
-            if (resultEntry.imageData) unresolved.imageData = resultEntry.imageData;
-          }
-        }
-      }
-      // Suppress this internal user message from the visible history
-      continue;
-    }
-
-    result.push({ ...msg, toolCalls: msg.toolCalls.map((tc) => ({ ...tc })) });
-  }
-  return result;
-}
-
-function handleUndo(
-  msg: UndoRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  const session = ctx.sessions.get(msg.sessionId);
-  if (!session) {
-    ctx.send(socket, { type: 'error', message: 'No active session' });
-    return;
-  }
-  const removedCount = session.undo();
-  ctx.send(socket, { type: 'undo_complete', removedCount, sessionId: msg.sessionId });
-}
-
-async function handleRegenerate(
-  msg: RegenerateRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): Promise<void> {
-  const session = ctx.sessions.get(msg.sessionId);
-  if (!session) {
-    ctx.send(socket, { type: 'error', message: 'No active session' });
-    return;
-  }
-
-  const sendEvent = (event: ServerMessage) => ctx.send(socket, event);
-  const requestId = uuid();
-  session.traceEmitter.emit('request_received', 'Regenerate requested', {
-    requestId,
-    status: 'info',
-    attributes: { source: 'regenerate' },
-  });
-  try {
-    await session.regenerate(sendEvent, requestId);
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err, sessionId: msg.sessionId }, 'Error regenerating message');
-    session.traceEmitter.emit('request_error', message.slice(0, 200), {
-      requestId,
-      status: 'error',
-      attributes: { errorClass: err instanceof Error ? err.constructor.name : 'Error', message: message.slice(0, 500) },
-    });
-    ctx.send(socket, { type: 'error', message: `Failed to regenerate: ${message}` });
-    const classified = classifySessionError(err, { phase: 'regenerate' });
-    ctx.send(socket, buildSessionErrorMessage(msg.sessionId, classified));
-  }
-}
-
-function handleUsageRequest(
-  msg: UsageRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  const conversation = conversationStore.getConversation(msg.sessionId);
-  if (!conversation) {
-    ctx.send(socket, { type: 'error', message: 'No active session' });
-    return;
-  }
-  const config = getConfig();
-  ctx.send(socket, {
-    type: 'usage_response',
-    totalInputTokens: conversation.totalInputTokens,
-    totalOutputTokens: conversation.totalOutputTokens,
-    estimatedCost: conversation.totalEstimatedCost,
-    model: config.model,
-  });
-}
-
-// ─── Skills handlers ─────────────────────────────────────────────────────────
-
-function handleSkillsList(socket: net.Socket, ctx: HandlerContext): void {
-  const config = getConfig();
-  const catalog = loadSkillCatalog();
-  const resolved = resolveSkillStates(catalog, config);
-
-  const skills = resolved.map((r) => ({
-    id: r.summary.id,
-    name: r.summary.name,
-    description: r.summary.description,
-    emoji: r.summary.emoji,
-    homepage: r.summary.homepage,
-    source: r.summary.source as 'bundled' | 'managed' | 'workspace' | 'clawhub' | 'extra',
-    state: (r.state === 'degraded' ? 'enabled' : r.state) as 'enabled' | 'disabled' | 'available',
-    degraded: r.degraded,
-    missingRequirements: r.missingRequirements,
-    updateAvailable: false,
-    userInvocable: r.summary.userInvocable,
-  }));
-
-  ctx.send(socket, { type: 'skills_list_response', skills });
-}
-
-/** Get or create the skill entry object for a given skill name, creating intermediate objects as needed.
- *  Guards against malformed config (e.g. skills or entries being a string, array, or null)
- *  by resetting non-object intermediates to {}, restoring self-healing behavior. */
-function ensureSkillEntry(raw: Record<string, unknown>, name: string): Record<string, unknown> {
-  if (!isRecord(raw.skills) || Array.isArray(raw.skills)) raw.skills = {};
-  const skills = raw.skills as Record<string, unknown>;
-  if (!isRecord(skills.entries) || Array.isArray(skills.entries)) skills.entries = {};
-  const entries = skills.entries as Record<string, unknown>;
-  if (!isRecord(entries[name]) || Array.isArray(entries[name])) entries[name] = {};
-  return entries[name] as Record<string, unknown>;
-}
-function handleSkillsEnable(
-  msg: SkillsEnableRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  try {
-    const raw = loadRawConfig();
-    ensureSkillEntry(raw, msg.name).enabled = true;
-
-    ctx.setSuppressConfigReload(true);
-    try {
-      saveRawConfig(raw);
-    } catch (err) {
-      ctx.setSuppressConfigReload(false);
-      throw err;
-    }
-    invalidateConfigCache();
-
-    const existingSuppressTimer = ctx.debounceTimers.get('__suppress_reset__');
-    if (existingSuppressTimer) clearTimeout(existingSuppressTimer);
-    const resetTimer = setTimeout(() => { ctx.setSuppressConfigReload(false); }, 300);
-    ctx.debounceTimers.set('__suppress_reset__', resetTimer);
-
-    ctx.updateConfigFingerprint();
-
-    ctx.send(socket, {
-      type: 'skills_operation_response',
-      operation: 'enable',
-      success: true,
-    });
-    ctx.broadcast({
-      type: 'skills_state_changed',
-      name: msg.name,
-      state: 'enabled',
-    });
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err }, 'Failed to enable skill');
-    ctx.send(socket, {
-      type: 'skills_operation_response',
-      operation: 'enable',
-      success: false,
-      error: message,
-    });
-  }
-}
-
-function handleSkillsDisable(
-  msg: SkillsDisableRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  try {
-    const raw = loadRawConfig();
-    ensureSkillEntry(raw, msg.name).enabled = false;
-
-    ctx.setSuppressConfigReload(true);
-    try {
-      saveRawConfig(raw);
-    } catch (err) {
-      ctx.setSuppressConfigReload(false);
-      throw err;
-    }
-    invalidateConfigCache();
-
-    const existingSuppressTimer = ctx.debounceTimers.get('__suppress_reset__');
-    if (existingSuppressTimer) clearTimeout(existingSuppressTimer);
-    const resetTimer = setTimeout(() => { ctx.setSuppressConfigReload(false); }, 300);
-    ctx.debounceTimers.set('__suppress_reset__', resetTimer);
-
-    ctx.updateConfigFingerprint();
-
-    ctx.send(socket, {
-      type: 'skills_operation_response',
-      operation: 'disable',
-      success: true,
-    });
-    ctx.broadcast({
-      type: 'skills_state_changed',
-      name: msg.name,
-      state: 'disabled',
-    });
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err }, 'Failed to disable skill');
-    ctx.send(socket, {
-      type: 'skills_operation_response',
-      operation: 'disable',
-      success: false,
-      error: message,
-    });
-  }
-}
-
-function handleSkillsConfigure(
-  msg: SkillsConfigureRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  try {
-    const raw = loadRawConfig();
-
-    const entry = ensureSkillEntry(raw, msg.name);
-    if (msg.env) {
-      entry.env = msg.env;
-    }
-    if (msg.apiKey !== undefined) {
-      entry.apiKey = msg.apiKey;
-    }
-    if (msg.config) {
-      entry.config = msg.config;
-    }
-
-    ctx.setSuppressConfigReload(true);
-    try {
-      saveRawConfig(raw);
-    } catch (err) {
-      ctx.setSuppressConfigReload(false);
-      throw err;
-    }
-    invalidateConfigCache();
-
-    const existingSuppressTimer = ctx.debounceTimers.get('__suppress_reset__');
-    if (existingSuppressTimer) clearTimeout(existingSuppressTimer);
-    const resetTimer = setTimeout(() => { ctx.setSuppressConfigReload(false); }, 300);
-    ctx.debounceTimers.set('__suppress_reset__', resetTimer);
-
-    ctx.updateConfigFingerprint();
-
-    ctx.send(socket, {
-      type: 'skills_operation_response',
-      operation: 'configure',
-      success: true,
-    });
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err }, 'Failed to configure skill');
-    ctx.send(socket, {
-      type: 'skills_operation_response',
-      operation: 'configure',
-      success: false,
-      error: message,
-    });
-  }
-}
-
-async function handleSkillsInstall(
-  msg: SkillsInstallRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): Promise<void> {
-  try {
-    const result = await clawhubInstall(msg.slug, { version: msg.version });
-    if (!result.success) {
-      ctx.send(socket, {
-        type: 'skills_operation_response',
-        operation: 'install',
-        success: false,
-        error: result.error ?? 'Unknown error',
-      });
-      return;
-    }
-
-    // Reload skill catalog so the newly installed skill is picked up
-    loadSkillCatalog();
-
-    ctx.send(socket, {
-      type: 'skills_operation_response',
-      operation: 'install',
-      success: true,
-    });
-    ctx.broadcast({
-      type: 'skills_state_changed',
-      name: result.skillName ?? msg.slug,
-      state: 'installed',
-    });
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err }, 'Failed to install skill');
-    ctx.send(socket, {
-      type: 'skills_operation_response',
-      operation: 'install',
-      success: false,
-      error: message,
-    });
-  }
-}
-
-async function handleSkillsUninstall(
-  msg: SkillsUninstallRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): Promise<void> {
-  // Validate skill name to prevent path traversal while allowing namespaced slugs (org/name)
-  const validNamespacedSlug = /^[a-zA-Z0-9][a-zA-Z0-9._-]*\/[a-zA-Z0-9][a-zA-Z0-9._-]*$/;
-  const validSimpleName = /^[a-zA-Z0-9][a-zA-Z0-9._-]*$/;
-  if (msg.name.includes('..') || msg.name.includes('\\') || !(validSimpleName.test(msg.name) || validNamespacedSlug.test(msg.name))) {
-    ctx.send(socket, {
-      type: 'skills_operation_response',
-      operation: 'uninstall',
-      success: false,
-      error: 'Invalid skill name',
-    });
-    return;
-  }
-
-  try {
-    // Use shared managed-store logic for simple managed skill IDs
-    const isManagedId = !validateManagedSkillId(msg.name);
-    if (isManagedId) {
-      const result = deleteManagedSkill(msg.name);
-      if (!result.deleted) {
-        ctx.send(socket, {
-          type: 'skills_operation_response',
-          operation: 'uninstall',
-          success: false,
-          error: result.error ?? 'Failed to delete managed skill',
-        });
-        return;
-      }
-    } else {
-      // Namespaced slug (org/name) — direct filesystem removal
-      const skillDir = join(getWorkspaceSkillsDir(), msg.name);
-      if (!existsSync(skillDir)) {
-        ctx.send(socket, {
-          type: 'skills_operation_response',
-          operation: 'uninstall',
-          success: false,
-          error: 'Skill not found',
-        });
-        return;
-      }
-      rmSync(skillDir, { recursive: true });
-      try { removeSkillsIndexEntry(msg.name); } catch { /* best effort */ }
-    }
-
-    // Clean config entry
-    const raw = loadRawConfig();
-    const skills = raw.skills as Record<string, unknown> | undefined;
-    const entries = skills?.entries as Record<string, unknown> | undefined;
-    if (entries?.[msg.name]) {
-      delete entries[msg.name];
-
-      ctx.setSuppressConfigReload(true);
-      try {
-        saveRawConfig(raw);
-      } catch (err) {
-        ctx.setSuppressConfigReload(false);
-        throw err;
-      }
-      invalidateConfigCache();
-
-      const existingSuppressTimer = ctx.debounceTimers.get('__suppress_reset__');
-      if (existingSuppressTimer) clearTimeout(existingSuppressTimer);
-      const resetTimer = setTimeout(() => { ctx.setSuppressConfigReload(false); }, 300);
-      ctx.debounceTimers.set('__suppress_reset__', resetTimer);
-
-      ctx.updateConfigFingerprint();
-    }
-
-    ctx.send(socket, {
-      type: 'skills_operation_response',
-      operation: 'uninstall',
-      success: true,
-    });
-    ctx.broadcast({
-      type: 'skills_state_changed',
-      name: msg.name,
-      state: 'uninstalled',
-    });
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err }, 'Failed to uninstall skill');
-    ctx.send(socket, {
-      type: 'skills_operation_response',
-      operation: 'uninstall',
-      success: false,
-      error: message,
-    });
-  }
-}
-
-async function handleSkillsUpdate(
-  msg: SkillsUpdateRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): Promise<void> {
-  try {
-    const result = await clawhubUpdate(msg.name);
-    if (!result.success) {
-      ctx.send(socket, {
-        type: 'skills_operation_response',
-        operation: 'update',
-        success: false,
-        error: result.error ?? 'Unknown error',
-      });
-      return;
-    }
-
-    // Reload skill catalog to pick up updated skill
-    loadSkillCatalog();
-
-    ctx.send(socket, {
-      type: 'skills_operation_response',
-      operation: 'update',
-      success: true,
-    });
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err }, 'Failed to update skill');
-    ctx.send(socket, {
-      type: 'skills_operation_response',
-      operation: 'update',
-      success: false,
-      error: message,
-    });
-  }
-}
-
-async function handleSkillsCheckUpdates(
-  _msg: SkillsCheckUpdatesRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): Promise<void> {
-  try {
-    const updates = await clawhubCheckUpdates();
-    ctx.send(socket, {
-      type: 'skills_operation_response',
-      operation: 'check_updates',
-      success: true,
-      data: updates,
-    });
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err }, 'Failed to check for skill updates');
-    ctx.send(socket, {
-      type: 'skills_operation_response',
-      operation: 'check_updates',
-      success: false,
-      error: message,
-    });
-  }
-}
-
-async function handleSkillsSearch(
-  msg: SkillsSearchRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): Promise<void> {
-  try {
-    const result = await clawhubSearch(msg.query);
-    ctx.send(socket, {
-      type: 'skills_operation_response',
-      operation: 'search',
-      success: true,
-      data: result,
-    });
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err }, 'Failed to search skills');
-    ctx.send(socket, {
-      type: 'skills_operation_response',
-      operation: 'search',
-      success: false,
-      error: message,
-    });
-  }
-}
-
-async function handleSkillsInspect(
-  msg: SkillsInspectRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): Promise<void> {
-  try {
-    const result = await clawhubInspect(msg.slug);
-    ctx.send(socket, {
-      type: 'skills_inspect_response',
-      slug: msg.slug,
-      ...(result.data ? { data: result.data } : {}),
-      ...(result.error ? { error: result.error } : {}),
-    });
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err }, 'Failed to inspect skill');
-    ctx.send(socket, {
-      type: 'skills_inspect_response',
-      slug: msg.slug,
-      error: message,
-    });
-  }
-}
-
-async function handleSkillDetail(
-  msg: SkillDetailRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): Promise<void> {
-  const result = loadSkillBySelector(msg.skillId);
-  if (result.skill) {
-    const icon = await ensureSkillIcon(result.skill.directoryPath, result.skill.name, result.skill.description);
-    ctx.send(socket, {
-      type: 'skill_detail_response',
-      skillId: result.skill.id,
-      body: result.skill.body,
-      ...(icon ? { icon } : {}),
-    });
-  } else {
-    ctx.send(socket, {
-      type: 'skill_detail_response',
-      skillId: msg.skillId,
-      body: '',
-      error: result.error ?? 'Skill not found',
-    });
-  }
-}
-
-// ─── App data handler ────────────────────────────────────────────────────────
-
-function handleAppDataRequest(
-  msg: AppDataRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  const { surfaceId, callId, method, appId, recordId, data } = msg;
-  try {
-    let result: unknown = null;
-    switch (method) {
-      case 'query':
-        result = queryAppRecords(appId);
-        break;
-      case 'create':
-        if (!data) throw new Error('data is required for create');
-        result = createAppRecord(appId, data);
-        break;
-      case 'update':
-        if (!recordId) throw new Error('recordId is required for update');
-        if (!data) throw new Error('data is required for update');
-        result = updateAppRecord(appId, recordId, data);
-        break;
-      case 'delete':
-        if (!recordId) throw new Error('recordId is required for delete');
-        deleteAppRecord(appId, recordId);
-        result = null;
-        break;
-      default:
-        throw new Error(`Unknown app data method: ${method}`);
-    }
-    ctx.send(socket, {
-      type: 'app_data_response',
-      surfaceId,
-      callId,
-      success: true,
-      result,
-    });
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err, method, appId, recordId }, 'Error handling app_data_request');
-    ctx.send(socket, {
-      type: 'app_data_response',
-      surfaceId,
-      callId,
-      success: false,
-      error: message,
-    });
-  }
-}
-
-// ─── Task submit handler ────────────────────────────────────────────────────
-
-async function handleTaskSubmit(
-  msg: TaskSubmit,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): Promise<void> {
-  const requestId = uuid();
-  const rlog = log.child({ requestId });
-
-  try {
-    // Block inbound tasks that contain secrets and redirect to secure prompt
-    const taskIngressCheck = checkIngressForSecrets(msg.task);
-    if (taskIngressCheck.blocked) {
-      rlog.warn({ detectedTypes: taskIngressCheck.detectedTypes }, 'Blocked task_submit containing secrets');
-      ctx.send(socket, {
-        type: 'error',
-        message: taskIngressCheck.userNotice!,
-      });
-      // Create an ephemeral session so the secret_response lifecycle works
-      // end-to-end. The conversation is deleted after the prompt resolves
-      // to avoid accumulating placeholder entries in session history.
-      const conversation = conversationStore.createConversation('(blocked — secret detected)');
-      ctx.socketToSession.set(socket, conversation.id);
-      const session = await ctx.getOrCreateSession(conversation.id, socket, true);
-      session.redirectToSecurePrompt(taskIngressCheck.detectedTypes, () => {
-        conversationStore.deleteConversation(conversation.id);
-        // Clean up in-memory session and socket binding so the ephemeral
-        // session doesn't accumulate in the daemon's session map.
-        const s = ctx.sessions.get(conversation.id);
-        if (s) {
-          s.dispose();
-          ctx.sessions.delete(conversation.id);
-        }
-        // Only unbind if the socket still points to this ephemeral conversation;
-        // a new task_submit may have already rebound it to a real session.
-        if (ctx.socketToSession.get(socket) === conversation.id) {
-          ctx.socketToSession.delete(socket);
-        }
-      });
-      return;
-    }
-
-    // Slash candidates always route to text_qa — bypass classifier
-    const slashCandidate = parseSlashCandidate(msg.task);
-    const interactionType = slashCandidate.kind === 'candidate'
-      ? 'text_qa' as const
-      : await classifyInteraction(msg.task, msg.source);
-    rlog.info({ interactionType, slashBypass: slashCandidate.kind === 'candidate', taskLength: msg.task.length }, 'Task classified');
-
-    if (interactionType === 'computer_use') {
-      // Create CU session (reuse handleCuSessionCreate logic)
-      const sessionId = uuid();
-      const cuMsg: CuSessionCreate = {
-        type: 'cu_session_create',
-        sessionId,
-        task: msg.task,
-        screenWidth: msg.screenWidth,
-        screenHeight: msg.screenHeight,
-        attachments: msg.attachments,
-        interactionType: 'computer_use',
-      };
-      handleCuSessionCreate(cuMsg, socket, ctx);
-
-      ctx.send(socket, {
-        type: 'task_routed',
-        sessionId,
-        interactionType: 'computer_use',
-      });
-    } else {
-      // Create text QA session and immediately start processing
-      const conversation = conversationStore.createConversation(msg.task);
-      ctx.socketToSession.set(socket, conversation.id);
-      const session = await ctx.getOrCreateSession(conversation.id, socket, true);
-
-      // Wire escalation handler so the agent can call request_computer_control
-      wireEscalationHandler(session, socket, ctx, msg.screenWidth, msg.screenHeight);
-
-      ctx.send(socket, {
-        type: 'task_routed',
-        sessionId: conversation.id,
-        interactionType: 'text_qa',
-      });
-
-      // Start streaming immediately — client doesn't need to send user_message
-      session.processMessage(msg.task, msg.attachments ?? [], (event) => {
-        ctx.send(socket, event);
-      }, requestId).catch((err) => {
-        const message = err instanceof Error ? err.message : String(err);
-        rlog.error({ err }, 'Error processing task_submit text QA');
-        ctx.send(socket, { type: 'error', message: `Failed to process message: ${message}` });
-        const classified = classifySessionError(err, { phase: 'agent_loop' });
-        ctx.send(socket, buildSessionErrorMessage(conversation.id, classified));
-      });
-    }
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    rlog.error({ err }, 'Error handling task_submit');
-    ctx.send(socket, { type: 'error', message: `Failed to route task: ${message}` });
-  }
-}
-
-// ─── Trust rule handler ─────────────────────────────────────────────────────
-
-function handleAddTrustRule(
-  msg: AddTrustRule,
-  _socket: net.Socket,
-  _ctx: HandlerContext,
-): void {
-  try {
-    addRule(msg.toolName, msg.pattern, msg.scope, msg.decision);
-    log.info({ tool: msg.toolName, pattern: msg.pattern, scope: msg.scope, decision: msg.decision }, 'Trust rule added via client');
-  } catch (err) {
-    log.error({ err }, 'Failed to add trust rule');
-  }
-}
-
-function handleTrustRulesList(socket: net.Socket, ctx: HandlerContext): void {
-  const rules = getAllRules();
-  ctx.send(socket, { type: 'trust_rules_list_response', rules });
-}
-
-function handleRemoveTrustRule(
-  msg: RemoveTrustRule,
-  _socket: net.Socket,
-  _ctx: HandlerContext,
-): void {
-  try {
-    const removed = removeRule(msg.id);
-    if (!removed) {
-      log.warn({ id: msg.id }, 'Trust rule not found for removal');
-    } else {
-      log.info({ id: msg.id }, 'Trust rule removed via client');
-    }
-  } catch (err) {
-    log.error({ err }, 'Failed to remove trust rule');
-  }
-}
-
-function handleUpdateTrustRule(
-  msg: UpdateTrustRule,
-  _socket: net.Socket,
-  _ctx: HandlerContext,
-): void {
-  try {
-    updateRule(msg.id, {
-      tool: msg.tool,
-      pattern: msg.pattern,
-      scope: msg.scope,
-      decision: msg.decision,
-      priority: msg.priority,
-    });
-    log.info({ id: msg.id }, 'Trust rule updated via client');
-  } catch (err) {
-    log.error({ err }, 'Failed to update trust rule');
-  }
-}
-
-// ─── Schedule handlers ──────────────────────────────────────────────────────
-
-function handleSchedulesList(socket: net.Socket, ctx: HandlerContext): void {
-  const jobs = listSchedules();
-  ctx.send(socket, {
-    type: 'schedules_list_response',
-    schedules: jobs.map((j) => ({
-      id: j.id,
-      name: j.name,
-      enabled: j.enabled,
-      cronExpression: j.cronExpression,
-      timezone: j.timezone,
-      message: j.message,
-      nextRunAt: j.nextRunAt,
-      lastRunAt: j.lastRunAt,
-      lastStatus: j.lastStatus,
-      description: describeCronExpression(j.cronExpression),
-    })),
-  });
-}
-
-function handleScheduleToggle(
-  msg: ScheduleToggle,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  try {
-    updateSchedule(msg.id, { enabled: msg.enabled });
-    log.info({ id: msg.id, enabled: msg.enabled }, 'Schedule toggled via client');
-  } catch (err) {
-    log.error({ err }, 'Failed to toggle schedule');
-  }
-  handleSchedulesList(socket, ctx);
-}
-
-function handleScheduleRemove(
-  msg: ScheduleRemove,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  try {
-    const removed = deleteSchedule(msg.id);
-    if (!removed) {
-      log.warn({ id: msg.id }, 'Schedule not found for removal');
-    } else {
-      log.info({ id: msg.id }, 'Schedule removed via client');
-    }
-  } catch (err) {
-    log.error({ err }, 'Failed to remove schedule');
-  }
-  handleSchedulesList(socket, ctx);
-}
-
-// ─── Reminder handlers ──────────────────────────────────────────────────────
-
-function handleRemindersList(socket: net.Socket, ctx: HandlerContext): void {
-  const items = listReminders();
-  ctx.send(socket, {
-    type: 'reminders_list_response',
-    reminders: items.map((r) => ({
-      id: r.id,
-      label: r.label,
-      message: r.message,
-      fireAt: r.fireAt,
-      mode: r.mode,
-      status: r.status,
-      firedAt: r.firedAt,
-      createdAt: r.createdAt,
-    })),
-  });
-}
-
-function handleReminderCancel(
-  msg: ReminderCancel,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  try {
-    const cancelled = cancelReminder(msg.id);
-    if (!cancelled) {
-      log.warn({ id: msg.id }, 'Reminder not found or already fired/cancelled');
-    } else {
-      log.info({ id: msg.id }, 'Reminder cancelled via client');
-    }
-  } catch (err) {
-    log.error({ err }, 'Failed to cancel reminder');
-  }
-  handleRemindersList(socket, ctx);
-}
-
-// ─── Suggestion handler ─────────────────────────────────────────────────────
-
-const SUGGESTION_CACHE_MAX = 100;
-const suggestionCache = new Map<string, string>();
-const suggestionInFlight = new Map<string, Promise<string | null>>();
-
-async function handleSuggestionRequest(
-  msg: SuggestionRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): Promise<void> {
-  const noSuggestion = () => {
-    ctx.send(socket, {
-      type: 'suggestion_response',
-      requestId: msg.requestId,
-      suggestion: null,
-      source: 'none' as const,
-    });
-  };
-
-  const rawMessages = conversationStore.getMessages(msg.sessionId);
-  if (rawMessages.length === 0) { noSuggestion(); return; }
-
-  // Find the most recent assistant message — only use it if it has text content.
-  // Do NOT fall back to older turns; if the latest assistant message is tool-only,
-  // return no suggestion rather than reusing stale text from a previous turn.
-  for (let i = rawMessages.length - 1; i >= 0; i--) {
-    const m = rawMessages[i];
-    if (m.role !== 'assistant') continue;
-
-    let content: unknown;
-    try { content = JSON.parse(m.content); } catch { content = m.content; }
-    const rendered = renderHistoryContent(content);
-    const text = rendered.text.trim();
-    if (!text) { noSuggestion(); return; }
-
-    // Return cached suggestion
-    const cached = suggestionCache.get(m.id);
-    if (cached !== undefined) {
-      ctx.send(socket, {
-        type: 'suggestion_response',
-        requestId: msg.requestId,
-        suggestion: cached,
-        source: 'llm' as const,
-      });
-      return;
-    }
-
-    // Try LLM suggestion if an Anthropic API key is configured
-    const apiKey = getConfig().apiKeys.anthropic;
-    if (apiKey) {
-      try {
-        let promise = suggestionInFlight.get(m.id);
-        if (!promise) {
-          promise = generateSuggestion(apiKey, text);
-          suggestionInFlight.set(m.id, promise);
-        }
-        const llmSuggestion = await promise;
-        suggestionInFlight.delete(m.id);
-
-        if (llmSuggestion) {
-          if (suggestionCache.size >= SUGGESTION_CACHE_MAX) {
-            const oldest = suggestionCache.keys().next().value!;
-            suggestionCache.delete(oldest);
-          }
-          suggestionCache.set(m.id, llmSuggestion);
-
-          ctx.send(socket, {
-            type: 'suggestion_response',
-            requestId: msg.requestId,
-            suggestion: llmSuggestion,
-            source: 'llm' as const,
-          });
-          return;
-        }
-      } catch (err) {
-        suggestionInFlight.delete(m.id);
-        log.warn({ err }, 'LLM suggestion failed');
-      }
-    }
-
-    noSuggestion();
-    return;
-  }
-
-  noSuggestion();
-}
-
-async function generateSuggestion(apiKey: string, assistantText: string): Promise<string | null> {
-  const client = new Anthropic({ apiKey });
-  const truncated = assistantText.length > 2000
-    ? assistantText.slice(-2000)
-    : assistantText;
-
-  const response = await client.messages.create({
-    model: 'claude-haiku-4-5-20251001',
-    max_tokens: 30,
-    messages: [
-      {
-        role: 'user',
-        content: `Given this assistant message, write a very short tab-complete suggestion (max 50 chars) the user could send next to keep the conversation going. Be casual, curious, or actionable — like a quick reply, not a formal request. Reply with ONLY the suggestion text.\n\nAssistant's message:\n${truncated}`,
-      },
-    ],
-  });
-
-  const textBlock = response.content.find((b) => b.type === 'text');
-  const raw = textBlock && 'text' in textBlock ? textBlock.text.trim() : '';
-  if (!raw || raw.length > 50) return null;
-
-  const firstLine = raw.split('\n')[0].trim();
-  return firstLine || null;
-}
-
-// ─── App open handler ───────────────────────────────────────────────────────
-
-function handleAppOpenRequest(msg: { appId: string }, socket: net.Socket, ctx: HandlerContext): void {
-  const appId = msg.appId;
-  if (!appId) {
-    ctx.send(socket, { type: 'error', message: 'app_open_request requires appId' });
-    return;
-  }
-
-  const app = getApp(appId);
-  if (!app) {
-    ctx.send(socket, { type: 'error', message: `App not found: ${appId}` });
-    return;
-  }
-
-  const surfaceId = `app-open-${uuid()}`;
-  ctx.send(socket, {
-    type: 'ui_surface_show',
-    sessionId: 'app-panel',
-    surfaceId,
-    surfaceType: 'dynamic_page',
-    title: app.name,
-    data: { html: app.htmlDefinition, appId: app.id, appType: app.appType },
-    display: 'panel',
-  } as UiSurfaceShow);
-}
-
-// ─── App update preview handler ─────────────────────────────────────────────
-
-function handleAppUpdatePreview(msg: AppUpdatePreviewRequest, socket: net.Socket, ctx: HandlerContext): void {
-  try {
-    updateApp(msg.appId, { preview: msg.preview });
-    ctx.send(socket, { type: 'app_update_preview_response', success: true, appId: msg.appId });
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err }, 'Failed to update app preview');
-    ctx.send(socket, { type: 'error', message: `Failed to update app preview: ${message}` });
-  }
-}
-
-// ─── Apps list handler ──────────────────────────────────────────────────────
-
-function handleAppsList(socket: net.Socket, ctx: HandlerContext): void {
-  try {
-    const apps = listApps();
-    ctx.send(socket, {
-      type: 'apps_list_response',
-      apps: apps.map((a) => {
-        const version = a.version ?? '1.0.0';
-        const contentId = computeContentId(a.name);
-        return {
-          id: a.id,
-          name: a.name,
-          description: a.description,
-          icon: a.icon,
-          preview: a.preview,
-          createdAt: a.createdAt,
-          version,
-          contentId,
-          appType: a.appType,
-        };
-      }),
-    });
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err }, 'Failed to list apps');
-    ctx.send(socket, { type: 'error', message: `Failed to list apps: ${message}` });
-  }
-}
-
-// ─── Shared apps handlers ────────────────────────────────────────────────────
-
-function getSharedAppsDir(): string {
-  return join(homedir(), 'Library', 'Application Support', 'vellum-assistant', 'shared-apps');
-}
-
-function handleSharedAppsList(socket: net.Socket, ctx: HandlerContext): void {
-  try {
-    const dir = getSharedAppsDir();
-    if (!existsSync(dir)) {
-      ctx.send(socket, { type: 'shared_apps_list_response', apps: [] });
-      return;
-    }
-
-    const files = readdirSync(dir).filter((f) => f.endsWith('-meta.json'));
-    const apps: Array<{
-      uuid: string;
-      name: string;
-      description?: string;
-      icon?: string;
-      preview?: string;
-      entry: string;
-      trustTier: string;
-      signerDisplayName?: string;
-      bundleSizeBytes: number;
-      installedAt: string;
-      version?: string;
-      contentId?: string;
-      forked?: boolean;
-    }> = [];
-
-    for (const file of files) {
-      try {
-        const raw = readFileSync(join(dir, file), 'utf-8');
-        const meta = JSON.parse(raw);
-
-        // Try to read version and content_id from the manifest.json inside the app dir
-        let version: string | undefined;
-        let contentId: string | undefined;
-        const manifestPath = join(dir, meta.uuid, 'manifest.json');
-        if (existsSync(manifestPath)) {
-          try {
-            const manifestRaw = readFileSync(manifestPath, 'utf-8');
-            const manifest = JSON.parse(manifestRaw);
-            version = manifest.version;
-            contentId = manifest.content_id;
-          } catch {
-            // ignore malformed manifest
-          }
-        }
-
-        apps.push({
-          uuid: meta.uuid,
-          name: meta.name,
-          description: meta.description,
-          icon: meta.icon,
-          preview: meta.preview,
-          entry: meta.entry,
-          trustTier: meta.trustTier,
-          signerDisplayName: meta.signerDisplayName,
-          bundleSizeBytes: meta.bundleSizeBytes ?? 0,
-          installedAt: meta.installedAt,
-          version,
-          contentId,
-          forked: meta.forked,
-        });
-      } catch {
-        log.warn({ file }, 'Failed to read shared app metadata file');
-      }
-    }
-
-    // Detect update availability for non-forked shared apps.
-    // Group by contentId, then mark older versions as having an update available.
-    const contentIdVersions = new Map<string, string[]>();
-    for (const app of apps) {
-      if (app.contentId && !app.forked) {
-        const versions = contentIdVersions.get(app.contentId) ?? [];
-        if (app.version) versions.push(app.version);
-        contentIdVersions.set(app.contentId, versions);
-      }
-    }
-
-    // Find the latest version for each contentId
-    const latestVersions = new Map<string, string>();
-    for (const [cid, versions] of contentIdVersions) {
-      if (versions.length > 0) {
-        versions.sort((a, b) => compareSemver(a, b));
-        latestVersions.set(cid, versions[versions.length - 1]);
-      }
-    }
-
-    const result = apps.map((app) => {
-      let updateAvailable = false;
-      if (app.contentId && app.version && !app.forked) {
-        const latest = latestVersions.get(app.contentId);
-        if (latest && compareSemver(app.version, latest) < 0) {
-          updateAvailable = true;
-        }
-      }
-      // Remove the internal forked field before sending
-      const { forked: _, ...rest } = app;
-      return { ...rest, updateAvailable: updateAvailable || undefined };
-    });
-
-    ctx.send(socket, { type: 'shared_apps_list_response', apps: result });
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err }, 'Failed to list shared apps');
-    ctx.send(socket, { type: 'error', message: `Failed to list shared apps: ${message}` });
-  }
-}
-
-/** Compare two semver strings. Returns negative if a < b, 0 if equal, positive if a > b. */
-function compareSemver(a: string, b: string): number {
-  const pa = a.split('.').map(Number);
-  const pb = b.split('.').map(Number);
-  for (let i = 0; i < 3; i++) {
-    const diff = (pa[i] ?? 0) - (pb[i] ?? 0);
-    if (diff !== 0) return diff;
-  }
-  return 0;
-}
-
-function handleSharedAppDelete(
-  msg: SharedAppDeleteRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  try {
-    const uuid = msg.uuid;
-    // Validate UUID to prevent path traversal
-    if (uuid.includes('/') || uuid.includes('\\') || uuid.includes('..')) {
-      ctx.send(socket, { type: 'shared_app_delete_response', success: false });
-      return;
-    }
-
-    const dir = getSharedAppsDir();
-    const appDir = join(dir, uuid);
-    const metaFile = join(dir, `${uuid}-meta.json`);
-
-    if (existsSync(appDir)) {
-      rmSync(appDir, { recursive: true });
-    }
-    if (existsSync(metaFile)) {
-      rmSync(metaFile);
-    }
-
-    ctx.send(socket, { type: 'shared_app_delete_response', success: true });
-  } catch (err) {
-    log.error({ err }, 'Failed to delete shared app');
-    ctx.send(socket, { type: 'shared_app_delete_response', success: false });
-  }
-}
-
-function handleForkSharedApp(
-  msg: ForkSharedAppRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  try {
-    const appUuid = msg.uuid;
-    // Validate UUID to prevent path traversal
-    if (appUuid.includes('/') || appUuid.includes('\\') || appUuid.includes('..') || /\s/.test(appUuid)) {
-      ctx.send(socket, { type: 'fork_shared_app_response', success: false, error: 'Invalid UUID' });
-      return;
-    }
-
-    const dir = getSharedAppsDir();
-    const metaFile = join(dir, `${appUuid}-meta.json`);
-
-    if (!existsSync(metaFile)) {
-      ctx.send(socket, { type: 'fork_shared_app_response', success: false, error: 'Shared app not found' });
-      return;
-    }
-
-    const metaRaw = readFileSync(metaFile, 'utf-8');
-    const meta = JSON.parse(metaRaw);
-    const appName = meta.name ?? 'Untitled';
-    const appDescription = meta.description;
-
-    // Read the HTML from the shared app's entry file
-    const entry = meta.entry ?? 'index.html';
-    const htmlPath = join(dir, appUuid, entry);
-
-    if (!existsSync(htmlPath)) {
-      ctx.send(socket, { type: 'fork_shared_app_response', success: false, error: 'Shared app HTML not found' });
-      return;
-    }
-
-    const htmlContent = readFileSync(htmlPath, 'utf-8');
-
-    // Create a new local app via the app store
-    const newApp = createApp({
-      name: `${appName} (Fork)`,
-      description: appDescription,
-      schemaJson: JSON.stringify({ type: 'object', properties: {} }),
-      htmlDefinition: htmlContent,
-    });
-
-    ctx.send(socket, {
-      type: 'fork_shared_app_response',
-      success: true,
-      appId: newApp.id,
-      name: newApp.name,
-    });
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err }, 'Failed to fork shared app');
-    ctx.send(socket, { type: 'fork_shared_app_response', success: false, error: message });
-  }
-}
-
-// ─── Share app cloud handler ────────────────────────────────────────────────
-
-async function handleShareAppCloud(
-  msg: ShareAppCloudRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): Promise<void> {
-  try {
-    const result = await packageApp(msg.appId);
-    const bundleData = readFileSync(result.bundlePath);
-    const { shareToken } = createSharedAppLink(bundleData, result.manifest);
-
-    const port = process.env.RUNTIME_HTTP_PORT
-      ? parseInt(process.env.RUNTIME_HTTP_PORT, 10)
-      : 7821;
-    const shareUrl = `http://localhost:${port}/v1/apps/shared/${shareToken}`;
-
-    ctx.send(socket, {
-      type: 'share_app_cloud_response',
-      success: true,
-      shareToken,
-      shareUrl,
-    });
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err, appId: msg.appId }, 'Failed to share app to cloud');
-    ctx.send(socket, {
-      type: 'share_app_cloud_response',
-      success: false,
-      error: message,
-    });
-  }
-}
-
-// ─── Publish page handlers ──────────────────────────────────────────────────
-
-async function handlePublishPage(
-  msg: PublishPageRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): Promise<void> {
-  try {
-    // Hash the HTML for dedup — can be done before credential check
-    const htmlHash = createHash('sha256').update(msg.html).digest('hex');
-
-    // Check if already published (no credential needed)
-    const existing = getPublishedPageByHash(htmlHash);
-    if (existing) {
-      ctx.send(socket, {
-        type: 'publish_page_response',
-        success: true,
-        publicUrl: existing.publicUrl,
-        deploymentId: existing.deploymentId,
-      });
-      return;
-    }
-
-    const publishExecute = async (token: string) => {
-      const name = msg.title
-        ? msg.title.toLowerCase().replace(/[^a-z0-9]+/g, '-').replace(/^-|-$/g, '').slice(0, 50)
-        : `vellum-page-${Date.now()}`;
-
-      const result = await deployHtmlToVercel({ html: msg.html, name, token });
-
-      const id = uuid();
-      createPublishedPage({
-        id,
-        deploymentId: result.deploymentId,
-        publicUrl: result.url,
-        pageTitle: msg.title,
-        htmlHash,
-      });
-
-      return { url: result.url, deploymentId: result.deploymentId };
-    };
-
-    let useResult = await credentialBroker.serverUse({
-      service: 'vercel',
-      field: 'api_token',
-      toolName: 'publish_page',
-      execute: publishExecute,
-    });
-
-    // If no credential found, prompt the user and retry
-    if (!useResult.success && useResult.reason?.includes('No credential found')) {
-      const allowedTools = ['publish_page', 'unpublish_page'];
-      const secretResult = await requestSecretStandalone(socket, ctx, {
-        service: 'vercel',
-        field: 'api_token',
-        label: 'Vercel API Token',
-        description: 'Required to publish site apps to the web. Create a token at vercel.com/account/tokens.',
-        placeholder: 'Enter your Vercel API token',
-        allowedTools,
-      });
-
-      if (!secretResult.value) {
-        ctx.send(socket, {
-          type: 'publish_page_response',
-          success: false,
-          error: 'Cancelled',
-        });
-        return;
-      }
-
-      // Store the credential
-      const storageKey = `credential:vercel:api_token`;
-      setSecureKey(storageKey, secretResult.value);
-      upsertCredentialMetadata('vercel', 'api_token', { allowedTools });
-
-      // Retry with the newly stored credential
-      useResult = await credentialBroker.serverUse({
-        service: 'vercel',
-        field: 'api_token',
-        toolName: 'publish_page',
-        execute: publishExecute,
-      });
-    }
-
-    if (useResult.success && useResult.result) {
-      ctx.send(socket, {
-        type: 'publish_page_response',
-        success: true,
-        publicUrl: useResult.result.url,
-        deploymentId: useResult.result.deploymentId,
-      });
-    } else {
-      log.error({ reason: useResult.reason }, 'Failed to publish page');
-      ctx.send(socket, {
-        type: 'publish_page_response',
-        success: false,
-        error: useResult.reason ?? 'Failed to publish page',
-      });
-    }
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err }, 'Failed to publish page');
-    ctx.send(socket, {
-      type: 'publish_page_response',
-      success: false,
-      error: message,
-    });
-  }
-}
-
-async function handleUnpublishPage(
-  msg: UnpublishPageRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): Promise<void> {
-  try {
-    const useResult = await credentialBroker.serverUse({
-      service: 'vercel',
-      field: 'api_token',
-      toolName: 'unpublish_page',
-      execute: async (token) => {
-        await deleteVercelDeployment(msg.deploymentId, token);
-
-        const record = getPublishedPageByDeploymentId(msg.deploymentId);
-        if (record) {
-          markDeleted(record.id);
-        }
-      },
-    });
-
-    if (useResult.success) {
-      ctx.send(socket, {
-        type: 'unpublish_page_response',
-        success: true,
-      });
-    } else {
-      log.error({ reason: useResult.reason, deploymentId: msg.deploymentId }, 'Failed to unpublish page');
-      ctx.send(socket, {
-        type: 'unpublish_page_response',
-        success: false,
-        error: useResult.reason ?? 'Failed to unpublish page',
-      });
-    }
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err, deploymentId: msg.deploymentId }, 'Failed to unpublish page');
-    ctx.send(socket, {
-      type: 'unpublish_page_response',
-      success: false,
-      error: message,
-    });
-  }
-}
-
-// ─── Bundle app handler ─────────────────────────────────────────────────────
-
-async function handleBundleApp(
-  msg: BundleAppRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): Promise<void> {
-  try {
-    const result = await packageApp(msg.appId);
-    ctx.send(socket, {
-      type: 'bundle_app_response',
-      bundlePath: result.bundlePath,
-      manifest: result.manifest,
-    });
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err, appId: msg.appId }, 'Failed to bundle app');
-    ctx.send(socket, { type: 'error', message: `Failed to bundle app: ${message}` });
-  }
-}
-
-// ─── IPC blob probe handler ─────────────────────────────────────────────────
-
-function handleIpcBlobProbe(
-  msg: IpcBlobProbe,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  if (!isValidBlobId(msg.probeId)) {
-    ctx.send(socket, {
-      type: 'ipc_blob_probe_result',
-      probeId: msg.probeId,
-      ok: false,
-      reason: 'invalid_probe_id',
-    });
-    return;
-  }
-
-  let filePath: string;
-  try {
-    filePath = resolveBlobPath(msg.probeId);
-  } catch {
-    ctx.send(socket, {
-      type: 'ipc_blob_probe_result',
-      probeId: msg.probeId,
-      ok: false,
-      reason: 'invalid_probe_id',
-    });
-    return;
-  }
-
-  let content: Buffer;
-  try {
-    content = readFileSync(filePath);
-  } catch {
-    ctx.send(socket, {
-      type: 'ipc_blob_probe_result',
-      probeId: msg.probeId,
-      ok: false,
-      reason: 'missing_probe_file',
-    });
-    return;
-  }
-
-  const observedHash = createHash('sha256').update(content).digest('hex');
-
-  // Best-effort cleanup regardless of match outcome
-  deleteBlob(msg.probeId);
-
-  if (observedHash !== msg.nonceSha256) {
-    ctx.send(socket, {
-      type: 'ipc_blob_probe_result',
-      probeId: msg.probeId,
-      ok: false,
-      observedNonceSha256: observedHash,
-      reason: 'hash_mismatch',
-    });
-    return;
-  }
-
-  ctx.send(socket, {
-    type: 'ipc_blob_probe_result',
-    probeId: msg.probeId,
-    ok: true,
-    observedNonceSha256: observedHash,
-  });
-}
-
-// ─── Gallery handlers ────────────────────────────────────────────────────────
-
-function handleGalleryList(socket: net.Socket, ctx: HandlerContext): void {
-  ctx.send(socket, { type: 'gallery_list_response', gallery: defaultGallery });
-}
-
-function handleGalleryInstall(
-  msg: GalleryInstallRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  try {
-    const galleryApp = defaultGallery.apps.find((a) => a.id === msg.galleryAppId);
-    if (!galleryApp) {
-      ctx.send(socket, {
-        type: 'gallery_install_response',
-        success: false,
-        error: `Gallery app not found: ${msg.galleryAppId}`,
-      });
-      return;
-    }
-
-    const app = createApp({
-      name: galleryApp.name,
-      description: galleryApp.description,
-      schemaJson: galleryApp.schemaJson,
-      htmlDefinition: galleryApp.htmlDefinition,
-    });
-
-    ctx.send(socket, {
-      type: 'gallery_install_response',
-      success: true,
-      appId: app.id,
-      name: app.name,
-    });
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err, galleryAppId: msg.galleryAppId }, 'Failed to install gallery app');
-    ctx.send(socket, {
-      type: 'gallery_install_response',
-      success: false,
-      error: message,
-    });
-  }
-}
-
-// ─── Slack handlers ─────────────────────────────────────────────────────────
-
-async function handleShareToSlack(
-  msg: ShareToSlackRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): Promise<void> {
-  try {
-    const config = loadRawConfig();
-    const webhookUrl = config.slackWebhookUrl as string | undefined;
-    if (!webhookUrl) {
-      ctx.send(socket, {
-        type: 'share_to_slack_response',
-        success: false,
-        error: 'No Slack webhook URL configured. Set one in Settings.',
-      });
-      return;
-    }
-
-    const app = getApp(msg.appId);
-    if (!app) {
-      ctx.send(socket, {
-        type: 'share_to_slack_response',
-        success: false,
-        error: `App not found: ${msg.appId}`,
-      });
-      return;
-    }
-
-    await postToSlackWebhook(
-      webhookUrl,
-      app.name,
-      app.description ?? '',
-      '\u{1F4F1}',
-    );
-
-    ctx.send(socket, { type: 'share_to_slack_response', success: true });
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err, appId: msg.appId }, 'Failed to share app to Slack');
-    ctx.send(socket, {
-      type: 'share_to_slack_response',
-      success: false,
-      error: message,
-    });
-  }
-}
-
-function handleSlackWebhookConfig(
-  msg: SlackWebhookConfigRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  try {
-    const config = loadRawConfig();
-    if (msg.action === 'get') {
-      ctx.send(socket, {
-        type: 'slack_webhook_config_response',
-        webhookUrl: (config.slackWebhookUrl as string) ?? undefined,
-        success: true,
-      });
-    } else {
-      config.slackWebhookUrl = msg.webhookUrl ?? '';
-      saveRawConfig(config);
-      ctx.send(socket, {
-        type: 'slack_webhook_config_response',
-        success: true,
-      });
-    }
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err }, 'Failed to handle Slack webhook config');
-    ctx.send(socket, {
-      type: 'slack_webhook_config_response',
-      success: false,
-      error: message,
-    });
-  }
-}
-
-function handleVercelApiConfig(
-  msg: VercelApiConfigRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  try {
-    if (msg.action === 'get') {
-      const existing = getSecureKey('credential:vercel:api_token');
-      ctx.send(socket, {
-        type: 'vercel_api_config_response',
-        hasToken: !!existing,
-        success: true,
-      });
-    } else if (msg.action === 'set') {
-      if (!msg.apiToken) {
-        ctx.send(socket, {
-          type: 'vercel_api_config_response',
-          hasToken: false,
-          success: false,
-          error: 'apiToken is required for set action',
-        });
-        return;
-      }
-      const stored = setSecureKey('credential:vercel:api_token', msg.apiToken);
-      if (!stored) {
-        ctx.send(socket, {
-          type: 'vercel_api_config_response',
-          hasToken: false,
-          success: false,
-          error: 'Failed to store API token in secure storage',
-        });
-        return;
-      }
-      upsertCredentialMetadata('vercel', 'api_token', {
-        allowedTools: ['publish_page', 'unpublish_page'],
-      });
-      ctx.send(socket, {
-        type: 'vercel_api_config_response',
-        hasToken: true,
-        success: true,
-      });
-    } else {
-      deleteSecureKey('credential:vercel:api_token');
-      deleteCredentialMetadata('vercel', 'api_token');
-      ctx.send(socket, {
-        type: 'vercel_api_config_response',
-        hasToken: false,
-        success: true,
-      });
-    }
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error({ err }, 'Failed to handle Vercel API config');
-    ctx.send(socket, {
-      type: 'vercel_api_config_response',
-      hasToken: false,
-      success: false,
-      error: message,
-    });
-  }
-}
-
-// ─── Computer-use handlers ──────────────────────────────────────────────────
-
-function removeCuSessionReferences(
-  ctx: HandlerContext,
-  sessionId: string,
-  expectedSession?: ComputerUseSession,
-): void {
-  const current = ctx.cuSessions.get(sessionId);
-  if (expectedSession && current && current !== expectedSession) {
-    return;
-  }
-  ctx.cuSessions.delete(sessionId);
-  cuObservationSequenceBySession.delete(sessionId);
-  ctx.cuObservationParseSequence.delete(sessionId);
-  for (const [sock, ids] of ctx.socketToCuSession) {
-    if (ids.delete(sessionId) && ids.size === 0) {
-      ctx.socketToCuSession.delete(sock);
-    }
-  }
-}
-
-function handleCuSessionCreate(
-  msg: CuSessionCreate,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  // Abort any existing session with the same ID to prevent zombies,
-  // and remove it from the previous owner's socket set so disconnect
-  // cleanup doesn't accidentally abort the replacement session.
-  const existingSession = ctx.cuSessions.get(msg.sessionId);
-  if (existingSession) {
-    existingSession.abort();
-    removeCuSessionReferences(ctx, msg.sessionId, existingSession);
-  }
-
-  const config = getConfig();
-  let provider = getFailoverProvider(config.provider, config.providerOrder);
-  const { rateLimit } = config;
-  if (rateLimit.maxRequestsPerMinute > 0 || rateLimit.maxTokensPerSession > 0) {
-    provider = new RateLimitProvider(provider, rateLimit, ctx.sharedRequestTimestamps);
-  }
-
-  const sendToClient = (serverMsg: ServerMessage) => {
-    ctx.send(socket, serverMsg);
-  };
-
-  const sessionRef: { current?: ComputerUseSession } = {};
-  const onTerminal = (sessionId: string) => {
-    removeCuSessionReferences(ctx, sessionId, sessionRef.current);
-    log.info({ sessionId }, 'Computer-use session cleaned up after terminal state');
-  };
-
-  const session = new ComputerUseSession(
-    msg.sessionId,
-    msg.task,
-    msg.screenWidth,
-    msg.screenHeight,
-    provider,
-    sendToClient,
-    msg.interactionType,
-    onTerminal,
-  );
-  sessionRef.current = session;
-
-  ctx.cuSessions.set(msg.sessionId, session);
-
-  // Track all CU sessions per socket so disconnect cleans up all of them
-  let sessionIds = ctx.socketToCuSession.get(socket);
-  if (!sessionIds) {
-    sessionIds = new Set();
-    ctx.socketToCuSession.set(socket, sessionIds);
-  }
-  sessionIds.add(msg.sessionId);
-
-  log.info({ sessionId: msg.sessionId, taskLength: msg.task.length }, 'Computer-use session created');
-}
-
-function handleCuSessionAbort(
-  msg: CuSessionAbort,
-  _socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  const session = ctx.cuSessions.get(msg.sessionId);
-  if (!session) {
-    log.debug({ sessionId: msg.sessionId }, 'CU session abort: session not found (already finished?)');
-    return;
-  }
-  session.abort();
-  removeCuSessionReferences(ctx, msg.sessionId, session);
-  log.info({ sessionId: msg.sessionId }, 'Computer-use session aborted by client');
-}
-
-async function handleCuObservation(
-  msg: CuObservation,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): Promise<void> {
-  const receiveTimestampMs = Date.now();
-
-  // Hydrate blob refs to inline values before any other processing.
-  // Strategy: blob-first, inline-fallback, cu_error if neither available.
-  if (msg.axTreeBlob) {
-    try {
-      validateBlobKindEncoding(msg.axTreeBlob, 'axTreeBlob');
-      const buf = await readBlob(msg.axTreeBlob);
-      msg.axTree = buf.toString('utf8');
-      deleteBlob(msg.axTreeBlob.id);
-    } catch (err) {
-      log.warn({ err, blobId: msg.axTreeBlob.id }, 'Failed to hydrate axTreeBlob, checking inline fallback');
-      deleteBlob(msg.axTreeBlob.id);
-      if (!msg.axTree) {
-        log.warn({ blobId: msg.axTreeBlob.id }, 'No inline axTree fallback; continuing with partial observation');
-      }
-    }
-  }
-
-  if (msg.screenshotBlob) {
-    try {
-      validateBlobKindEncoding(msg.screenshotBlob, 'screenshotBlob');
-      const buf = await readBlob(msg.screenshotBlob);
-      msg.screenshot = buf.toString('base64');
-      deleteBlob(msg.screenshotBlob.id);
-    } catch (err) {
-      log.warn({ err, blobId: msg.screenshotBlob.id }, 'Failed to hydrate screenshotBlob, checking inline fallback');
-      deleteBlob(msg.screenshotBlob.id);
-      if (!msg.screenshot) {
-        log.warn({ blobId: msg.screenshotBlob.id }, 'No inline screenshot fallback; continuing with partial observation');
-      }
-    }
-  }
-
-  const previousSequence = cuObservationSequenceBySession.get(msg.sessionId) ?? 0;
-  const sequence = previousSequence + 1;
-  cuObservationSequenceBySession.set(msg.sessionId, sequence);
-  const axTreeBytes = msg.axTree ? Buffer.byteLength(msg.axTree, 'utf8') : 0;
-  const axDiffBytes = msg.axDiff ? Buffer.byteLength(msg.axDiff, 'utf8') : 0;
-  const secondaryWindowsBytes = msg.secondaryWindows ? Buffer.byteLength(msg.secondaryWindows, 'utf8') : 0;
-  const screenshotBase64Bytes = msg.screenshot ? Buffer.byteLength(msg.screenshot, 'utf8') : 0;
-  const screenshotApproxRawBytes = msg.screenshot
-    ? Math.floor((msg.screenshot.length / 4) * 3)
-    : 0;
-  log.info({
-    sessionId: msg.sessionId,
-    sequence,
-    receiveTimestampMs,
-    axTreeBytes,
-    axDiffBytes,
-    secondaryWindowsBytes,
-    screenshotBase64Bytes,
-    screenshotApproxRawBytes,
-  }, 'IPC_METRIC cu_observation_daemon_receive');
-
-  const session = ctx.cuSessions.get(msg.sessionId);
-  if (!session) {
-    ctx.send(socket, {
-      type: 'cu_error',
-      sessionId: msg.sessionId,
-      message: `No computer-use session found for id ${msg.sessionId}`,
-    });
-    return;
-  }
-
-  // Fire-and-forget: the session sends messages via its sendToClient callback
-  session.handleObservation(msg).catch((err) => {
-    log.error({ err, sessionId: msg.sessionId }, 'Error handling CU observation');
-  });
-}
-
-function handleLinkOpenRequest(
-  msg: LinkOpenRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  try {
-    const parsed = new URL(msg.url);
-    if (!['http:', 'https:'].includes(parsed.protocol)) {
-      log.warn({ url: msg.url }, 'link_open_request: blocked non-http URL');
-      return;
-    }
-  } catch {
-    log.warn({ url: msg.url }, 'link_open_request: invalid URL');
-    return;
-  }
-  // V1: passthrough. Future: affiliate param injection based on metadata
-  const finalUrl = msg.url;
-  ctx.send(socket, { type: 'open_url', url: finalUrl });
-}
-
-// ─── Integration handlers ────────────────────────────────────────────────────
-
-function handleIntegrationList(
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  const statuses = listStatuses();
-  ctx.send(socket, {
-    type: 'integration_list_response',
-    integrations: statuses.map((s) => ({
-      id: s.id,
-      connected: s.connected,
-      accountInfo: s.accountInfo,
-      connectedAt: s.connectedAt,
-      lastUsed: s.lastUsed,
-      error: s.error,
-    })),
-  });
-}
-
-async function handleIntegrationConnect(
-  msg: IntegrationConnectRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): Promise<void> {
-  const def = getIntegration(msg.integrationId);
-  if (!def) {
-    ctx.send(socket, {
-      type: 'integration_connect_result',
-      integrationId: msg.integrationId,
-      success: false,
-      error: `Integration "${msg.integrationId}" not found`,
-    });
-    return;
-  }
-
-  if (def.authType !== 'oauth2' || !def.oauth2Config) {
-    ctx.send(socket, {
-      type: 'integration_connect_result',
-      integrationId: msg.integrationId,
-      success: false,
-      error: `Integration "${msg.integrationId}" does not support OAuth2`,
-    });
-    return;
-  }
-
-  // Load clientId from config
-  const config = getConfig();
-  const integrationConfig = config.integrations[msg.integrationId];
-  const clientId = integrationConfig?.clientId || def.oauth2Config.clientId;
-
-  if (!clientId) {
-    ctx.send(socket, {
-      type: 'integration_connect_result',
-      integrationId: msg.integrationId,
-      success: false,
-      error: `No clientId configured for "${msg.integrationId}". Set integrations.${msg.integrationId}.clientId in config.`,
-    });
-    return;
-  }
-
-  try {
-    const oauthConfig = { ...def.oauth2Config, clientId };
-    const { tokens, grantedScopes } = await startOAuth2Flow(oauthConfig, {
-      openUrl: (url) => ctx.send(socket, { type: 'open_url', url, title: `Connect ${def.name}` }),
-    });
-
-    // Store tokens in vault
-    const tokenStored = setSecureKey(`integration:${def.id}:access_token`, tokens.accessToken);
-    if (!tokenStored) {
-      ctx.send(socket, {
-        type: 'integration_connect_result',
-        integrationId: msg.integrationId,
-        success: false,
-        error: 'Failed to store access token in secure storage',
-      });
-      return;
-    }
-
-    const expiresAt = tokens.expiresIn ? Date.now() + tokens.expiresIn * 1000 : undefined;
-
-    // Fetch account info (email) if the integration has a userinfo endpoint
-    let accountInfo: string | undefined;
-    const userinfoUrl = def.oauth2Config.userinfoUrl;
-    if (userinfoUrl && grantedScopes.some((s) => s.includes('userinfo'))) {
-      try {
-        const resp = await fetch(userinfoUrl, {
-          headers: { Authorization: `Bearer ${tokens.accessToken}` },
-        });
-        if (resp.ok) {
-          const info = await resp.json() as { email?: string };
-          accountInfo = info.email;
-        }
-      } catch {
-        // Non-fatal — account info is optional
-      }
-    }
-
-    upsertCredentialMetadata(`integration:${def.id}`, 'access_token', {
-      allowedTools: def.allowedTools,
-      expiresAt,
-      grantedScopes,
-      accountInfo: accountInfo ?? null,
-    });
-
-    if (tokens.refreshToken) {
-      const refreshStored = setSecureKey(`integration:${def.id}:refresh_token`, tokens.refreshToken);
-      if (!refreshStored) {
-        log.warn({ integrationId: def.id }, 'Failed to store refresh token in secure storage');
-      } else {
-        upsertCredentialMetadata(`integration:${def.id}`, 'refresh_token', {});
-      }
-    }
-
-    ctx.send(socket, {
-      type: 'integration_connect_result',
-      integrationId: def.id,
-      success: true,
-      accountInfo,
-    });
-  } catch (err: unknown) {
-    const message = err instanceof Error ? err.message : 'Unknown error during OAuth flow';
-    log.error({ err, integrationId: msg.integrationId }, 'Integration connect failed');
-    ctx.send(socket, {
-      type: 'integration_connect_result',
-      integrationId: msg.integrationId,
-      success: false,
-      error: message,
-    });
-  }
-}
-
-function handleIntegrationDisconnect(
-  msg: IntegrationDisconnectRequest,
-  socket: net.Socket,
-  ctx: HandlerContext,
-): void {
-  disconnectIntegration(msg.integrationId);
-  log.info({ integrationId: msg.integrationId }, 'Integration disconnected');
-  // Send updated list so the client refreshes
-  handleIntegrationList(socket, ctx);
-}
+// Re-export everything from the decomposed handler modules.
+// This file exists for backwards compatibility — all imports from
+// './handlers.js' continue to work without changes.
+export {
+  handleMessage,
+  renderHistoryContent,
+  mergeToolResults,
+} from './handlers/index.js';
+
+export type {
+  HandlerContext,
+  SessionCreateOptions,
+  HistoryToolCall,
+  HistorySurface,
+  RenderedHistoryContent,
+  ParsedHistoryMessage,
+} from './handlers/index.js';
