+++ assistant/src/daemon/handlers/apps.ts
@@ -1,0 +1,428 @@
+import * as net from 'node:net';
+import { v4 as uuid } from 'uuid';
+import { existsSync, rmSync, readdirSync, readFileSync } from 'node:fs';
+import { homedir } from 'node:os';
+import { join } from 'node:path';
+import { queryAppRecords, createAppRecord, updateAppRecord, deleteAppRecord, listApps, getApp, createApp, updateApp } from '../../memory/app-store.js';
+import { computeContentId } from '../../util/content-id.js';
+import { packageApp } from '../../bundler/app-bundler.js';
+import { createSharedAppLink } from '../../memory/shared-app-links-store.js';
+import { defaultGallery } from '../../gallery/default-gallery.js';
+import type {
+  AppDataRequest,
+  BundleAppRequest,
+  SharedAppDeleteRequest,
+  ForkSharedAppRequest,
+  ShareAppCloudRequest,
+  GalleryInstallRequest,
+  AppUpdatePreviewRequest,
+  UiSurfaceShow,
+} from '../ipc-protocol.js';
+import { log, isRecord, compareSemver, type HandlerContext } from './shared.js';
+
+export function handleAppDataRequest(
+  msg: AppDataRequest,
+  socket: net.Socket,
+  ctx: HandlerContext,
+): void {
+  const { surfaceId, callId, method, appId, recordId, data } = msg;
+  try {
+    let result: unknown = null;
+    switch (method) {
+      case 'query':
+        result = queryAppRecords(appId);
+        break;
+      case 'create':
+        if (!data) throw new Error('data is required for create');
+        result = createAppRecord(appId, data);
+        break;
+      case 'update':
+        if (!recordId) throw new Error('recordId is required for update');
+        if (!data) throw new Error('data is required for update');
+        result = updateAppRecord(appId, recordId, data);
+        break;
+      case 'delete':
+        if (!recordId) throw new Error('recordId is required for delete');
+        deleteAppRecord(appId, recordId);
+        result = null;
+        break;
+      default:
+        throw new Error(`Unknown app data method: ${method}`);
+    }
+    ctx.send(socket, {
+      type: 'app_data_response',
+      surfaceId,
+      callId,
+      success: true,
+      result,
+    });
+  } catch (err) {
+    const message = err instanceof Error ? err.message : String(err);
+    log.error({ err, method, appId, recordId }, 'Error handling app_data_request');
+    ctx.send(socket, {
+      type: 'app_data_response',
+      surfaceId,
+      callId,
+      success: false,
+      error: message,
+    });
+  }
+}
+
+export function handleAppOpenRequest(msg: { appId: string }, socket: net.Socket, ctx: HandlerContext): void {
+  const appId = msg.appId;
+  if (!appId) {
+    ctx.send(socket, { type: 'error', message: 'app_open_request requires appId' });
+    return;
+  }
+
+  const app = getApp(appId);
+  if (!app) {
+    ctx.send(socket, { type: 'error', message: `App not found: ${appId}` });
+    return;
+  }
+
+  const surfaceId = `app-open-${uuid()}`;
+  ctx.send(socket, {
+    type: 'ui_surface_show',
+    sessionId: 'app-panel',
+    surfaceId,
+    surfaceType: 'dynamic_page',
+    title: app.name,
+    data: { html: app.htmlDefinition, appId: app.id, appType: app.appType },
+    display: 'panel',
+  } as UiSurfaceShow);
+}
+
+export function handleAppUpdatePreview(msg: AppUpdatePreviewRequest, socket: net.Socket, ctx: HandlerContext): void {
+  try {
+    updateApp(msg.appId, { preview: msg.preview });
+    ctx.send(socket, { type: 'app_update_preview_response', success: true, appId: msg.appId });
+  } catch (err) {
+    const message = err instanceof Error ? err.message : String(err);
+    log.error({ err }, 'Failed to update app preview');
+    ctx.send(socket, { type: 'error', message: `Failed to update app preview: ${message}` });
+  }
+}
+
+export function handleAppsList(socket: net.Socket, ctx: HandlerContext): void {
+  try {
+    const apps = listApps();
+    ctx.send(socket, {
+      type: 'apps_list_response',
+      apps: apps.map((a) => {
+        const version = a.version ?? '1.0.0';
+        const contentId = computeContentId(a.name);
+        return {
+          id: a.id,
+          name: a.name,
+          description: a.description,
+          icon: a.icon,
+          preview: a.preview,
+          createdAt: a.createdAt,
+          version,
+          contentId,
+          appType: a.appType,
+        };
+      }),
+    });
+  } catch (err) {
+    const message = err instanceof Error ? err.message : String(err);
+    log.error({ err }, 'Failed to list apps');
+    ctx.send(socket, { type: 'error', message: `Failed to list apps: ${message}` });
+  }
+}
+
+function getSharedAppsDir(): string {
+  return join(homedir(), 'Library', 'Application Support', 'vellum-assistant', 'shared-apps');
+}
+
+export function handleSharedAppsList(socket: net.Socket, ctx: HandlerContext): void {
+  try {
+    const dir = getSharedAppsDir();
+    if (!existsSync(dir)) {
+      ctx.send(socket, { type: 'shared_apps_list_response', apps: [] });
+      return;
+    }
+
+    const files = readdirSync(dir).filter((f) => f.endsWith('-meta.json'));
+    const apps: Array<{
+      uuid: string;
+      name: string;
+      description?: string;
+      icon?: string;
+      preview?: string;
+      entry: string;
+      trustTier: string;
+      signerDisplayName?: string;
+      bundleSizeBytes: number;
+      installedAt: string;
+      version?: string;
+      contentId?: string;
+      forked?: boolean;
+    }> = [];
+
+    for (const file of files) {
+      try {
+        const raw = readFileSync(join(dir, file), 'utf-8');
+        const meta = JSON.parse(raw);
+
+        // Try to read version and content_id from the manifest.json inside the app dir
+        let version: string | undefined;
+        let contentId: string | undefined;
+        const manifestPath = join(dir, meta.uuid, 'manifest.json');
+        if (existsSync(manifestPath)) {
+          try {
+            const manifestRaw = readFileSync(manifestPath, 'utf-8');
+            const manifest = JSON.parse(manifestRaw);
+            version = manifest.version;
+            contentId = manifest.content_id;
+          } catch {
+            // ignore malformed manifest
+          }
+        }
+
+        apps.push({
+          uuid: meta.uuid,
+          name: meta.name,
+          description: meta.description,
+          icon: meta.icon,
+          preview: meta.preview,
+          entry: meta.entry,
+          trustTier: meta.trustTier,
+          signerDisplayName: meta.signerDisplayName,
+          bundleSizeBytes: meta.bundleSizeBytes ?? 0,
+          installedAt: meta.installedAt,
+          version,
+          contentId,
+          forked: meta.forked,
+        });
+      } catch {
+        log.warn({ file }, 'Failed to read shared app metadata file');
+      }
+    }
+
+    // Detect update availability for non-forked shared apps.
+    // Group by contentId, then mark older versions as having an update available.
+    const contentIdVersions = new Map<string, string[]>();
+    for (const app of apps) {
+      if (app.contentId && !app.forked) {
+        const versions = contentIdVersions.get(app.contentId) ?? [];
+        if (app.version) versions.push(app.version);
+        contentIdVersions.set(app.contentId, versions);
+      }
+    }
+
+    // Find the latest version for each contentId
+    const latestVersions = new Map<string, string>();
+    for (const [cid, versions] of contentIdVersions) {
+      if (versions.length > 0) {
+        versions.sort((a, b) => compareSemver(a, b));
+        latestVersions.set(cid, versions[versions.length - 1]);
+      }
+    }
+
+    const result = apps.map((app) => {
+      let updateAvailable = false;
+      if (app.contentId && app.version && !app.forked) {
+        const latest = latestVersions.get(app.contentId);
+        if (latest && compareSemver(app.version, latest) < 0) {
+          updateAvailable = true;
+        }
+      }
+      // Remove the internal forked field before sending
+      const { forked: _, ...rest } = app;
+      return { ...rest, updateAvailable: updateAvailable || undefined };
+    });
+
+    ctx.send(socket, { type: 'shared_apps_list_response', apps: result });
+  } catch (err) {
+    const message = err instanceof Error ? err.message : String(err);
+    log.error({ err }, 'Failed to list shared apps');
+    ctx.send(socket, { type: 'error', message: `Failed to list shared apps: ${message}` });
+  }
+}
+
+export function handleSharedAppDelete(
+  msg: SharedAppDeleteRequest,
+  socket: net.Socket,
+  ctx: HandlerContext,
+): void {
+  try {
+    const uuid = msg.uuid;
+    // Validate UUID to prevent path traversal
+    if (uuid.includes('/') || uuid.includes('\\') || uuid.includes('..')) {
+      ctx.send(socket, { type: 'shared_app_delete_response', success: false });
+      return;
+    }
+
+    const dir = getSharedAppsDir();
+    const appDir = join(dir, uuid);
+    const metaFile = join(dir, `${uuid}-meta.json`);
+
+    if (existsSync(appDir)) {
+      rmSync(appDir, { recursive: true });
+    }
+    if (existsSync(metaFile)) {
+      rmSync(metaFile);
+    }
+
+    ctx.send(socket, { type: 'shared_app_delete_response', success: true });
+  } catch (err) {
+    log.error({ err }, 'Failed to delete shared app');
+    ctx.send(socket, { type: 'shared_app_delete_response', success: false });
+  }
+}
+
+export function handleForkSharedApp(
+  msg: ForkSharedAppRequest,
+  socket: net.Socket,
+  ctx: HandlerContext,
+): void {
+  try {
+    const appUuid = msg.uuid;
+    // Validate UUID to prevent path traversal
+    if (appUuid.includes('/') || appUuid.includes('\\') || appUuid.includes('..') || /\s/.test(appUuid)) {
+      ctx.send(socket, { type: 'fork_shared_app_response', success: false, error: 'Invalid UUID' });
+      return;
+    }
+
+    const dir = getSharedAppsDir();
+    const metaFile = join(dir, `${appUuid}-meta.json`);
+
+    if (!existsSync(metaFile)) {
+      ctx.send(socket, { type: 'fork_shared_app_response', success: false, error: 'Shared app not found' });
+      return;
+    }
+
+    const metaRaw = readFileSync(metaFile, 'utf-8');
+    const meta = JSON.parse(metaRaw);
+    const appName = meta.name ?? 'Untitled';
+    const appDescription = meta.description;
+
+    // Read the HTML from the shared app's entry file
+    const entry = meta.entry ?? 'index.html';
+    const htmlPath = join(dir, appUuid, entry);
+
+    if (!existsSync(htmlPath)) {
+      ctx.send(socket, { type: 'fork_shared_app_response', success: false, error: 'Shared app HTML not found' });
+      return;
+    }
+
+    const htmlContent = readFileSync(htmlPath, 'utf-8');
+
+    // Create a new local app via the app store
+    const newApp = createApp({
+      name: `${appName} (Fork)`,
+      description: appDescription,
+      schemaJson: JSON.stringify({ type: 'object', properties: {} }),
+      htmlDefinition: htmlContent,
+    });
+
+    ctx.send(socket, {
+      type: 'fork_shared_app_response',
+      success: true,
+      appId: newApp.id,
+      name: newApp.name,
+    });
+  } catch (err) {
+    const message = err instanceof Error ? err.message : String(err);
+    log.error({ err }, 'Failed to fork shared app');
+    ctx.send(socket, { type: 'fork_shared_app_response', success: false, error: message });
+  }
+}
+
+export async function handleShareAppCloud(
+  msg: ShareAppCloudRequest,
+  socket: net.Socket,
+  ctx: HandlerContext,
+): Promise<void> {
+  try {
+    const result = await packageApp(msg.appId);
+    const bundleData = readFileSync(result.bundlePath);
+    const { shareToken } = createSharedAppLink(bundleData, result.manifest);
+
+    const port = process.env.RUNTIME_HTTP_PORT
+      ? parseInt(process.env.RUNTIME_HTTP_PORT, 10)
+      : 7821;
+    const shareUrl = `http://localhost:${port}/v1/apps/shared/${shareToken}`;
+
+    ctx.send(socket, {
+      type: 'share_app_cloud_response',
+      success: true,
+      shareToken,
+      shareUrl,
+    });
+  } catch (err) {
+    const message = err instanceof Error ? err.message : String(err);
+    log.error({ err, appId: msg.appId }, 'Failed to share app to cloud');
+    ctx.send(socket, {
+      type: 'share_app_cloud_response',
+      success: false,
+      error: message,
+    });
+  }
+}
+
+export async function handleBundleApp(
+  msg: BundleAppRequest,
+  socket: net.Socket,
+  ctx: HandlerContext,
+): Promise<void> {
+  try {
+    const result = await packageApp(msg.appId);
+    ctx.send(socket, {
+      type: 'bundle_app_response',
+      bundlePath: result.bundlePath,
+      manifest: result.manifest,
+    });
+  } catch (err) {
+    const message = err instanceof Error ? err.message : String(err);
+    log.error({ err, appId: msg.appId }, 'Failed to bundle app');
+    ctx.send(socket, { type: 'error', message: `Failed to bundle app: ${message}` });
+  }
+}
+
+export function handleGalleryList(socket: net.Socket, ctx: HandlerContext): void {
+  ctx.send(socket, { type: 'gallery_list_response', gallery: defaultGallery });
+}
+
+export function handleGalleryInstall(
+  msg: GalleryInstallRequest,
+  socket: net.Socket,
+  ctx: HandlerContext,
+): void {
+  try {
+    const galleryApp = defaultGallery.apps.find((a) => a.id === msg.galleryAppId);
+    if (!galleryApp) {
+      ctx.send(socket, {
+        type: 'gallery_install_response',
+        success: false,
+        error: `Gallery app not found: ${msg.galleryAppId}`,
+      });
+      return;
+    }
+
+    const app = createApp({
+      name: galleryApp.name,
+      description: galleryApp.description,
+      schemaJson: galleryApp.schemaJson,
+      htmlDefinition: galleryApp.htmlDefinition,
+    });
+
+    ctx.send(socket, {
+      type: 'gallery_install_response',
+      success: true,
+      appId: app.id,
+      name: app.name,
+    });
+  } catch (err) {
+    const message = err instanceof Error ? err.message : String(err);
+    log.error({ err, galleryAppId: msg.galleryAppId }, 'Failed to install gallery app');
+    ctx.send(socket, {
+      type: 'gallery_install_response',
+      success: false,
+      error: message,
+    });
+  }
+}
