+++ assistant/src/daemon/handlers/shared.ts
@@ -1,0 +1,469 @@
+import * as net from 'node:net';
+import { v4 as uuid } from 'uuid';
+import { Session } from '../session.js';
+import { ComputerUseSession } from '../computer-use-session.js';
+import { getLogger } from '../../util/logger.js';
+import { execSync } from 'node:child_process';
+import { estimateBase64Bytes } from '../assistant-attachments.js';
+import type { CuSessionCreate, ServerMessage } from '../ipc-protocol.js';
+import type { SecretPromptResult } from '../../permissions/secret-prompter.js';
+
+const log = getLogger('handlers');
+
+export { log };
+
+const HISTORY_ATTACHMENT_TEXT_LIMIT = 500;
+
+export const FALLBACK_SCREEN = { width: 1920, height: 1080 };
+let cachedScreenDims: { width: number; height: number } | null = null;
+
+// Module-level map for non-session secret prompts (e.g. publish_page)
+export const pendingStandaloneSecrets = new Map<string, { resolve: (result: SecretPromptResult) => void; timer: ReturnType<typeof setTimeout> }>();
+
+export interface HistoryToolCall {
+  name: string;
+  input: Record<string, unknown>;
+  result?: string;
+  isError?: boolean;
+  /** Base64-encoded image data from tool contentBlocks (e.g. browser_screenshot). */
+  imageData?: string;
+}
+
+export interface HistorySurface {
+  surfaceId: string;
+  surfaceType: string;
+  title?: string;
+  data: Record<string, unknown>;
+  actions?: Array<{ id: string; label: string; style?: string }>;
+  display?: string;
+}
+
+export interface RenderedHistoryContent {
+  text: string;
+  toolCalls: HistoryToolCall[];
+  /** True when the first tool_use block appeared before any text block. */
+  toolCallsBeforeText: boolean;
+  /** Text segments split by tool-call boundaries. */
+  textSegments: string[];
+  /** Content block ordering using "text:N", "tool:N", "surface:N" encoding. */
+  contentOrder: string[];
+  /** UI surfaces (widgets) embedded in the message. */
+  surfaces: HistorySurface[];
+}
+
+export interface ParsedHistoryMessage {
+  id?: string;
+  role: string;
+  text: string;
+  timestamp: number;
+  toolCalls: HistoryToolCall[];
+  toolCallsBeforeText: boolean;
+  textSegments: string[];
+  contentOrder: string[];
+  surfaces: HistorySurface[];
+}
+
+/**
+ * Optional overrides for session creation (e.g. interview mode).
+ */
+export interface SessionCreateOptions {
+  systemPromptOverride?: string;
+  maxResponseTokens?: number;
+}
+
+/**
+ * Shared context that handlers need from the DaemonServer.
+ * Keeps handlers decoupled from the server class itself.
+ */
+export interface HandlerContext {
+  sessions: Map<string, Session>;
+  socketToSession: Map<net.Socket, string>;
+  cuSessions: Map<string, ComputerUseSession>;
+  socketToCuSession: Map<net.Socket, Set<string>>;
+  cuObservationParseSequence: Map<string, number>;
+  socketSandboxOverride: Map<net.Socket, boolean>;
+  sharedRequestTimestamps: number[];
+  debounceTimers: Map<string, ReturnType<typeof setTimeout>>;
+  suppressConfigReload: boolean;
+  setSuppressConfigReload(value: boolean): void;
+  updateConfigFingerprint(): void;
+  send(socket: net.Socket, msg: ServerMessage): void;
+  broadcast(msg: ServerMessage): void;
+  clearAllSessions(): number;
+  getOrCreateSession(
+    conversationId: string,
+    socket?: net.Socket,
+    rebindClient?: boolean,
+    options?: SessionCreateOptions,
+  ): Promise<Session>;
+}
+
+/**
+ * Query the main display dimensions via CoreGraphics.
+ * Cached after the first successful call; falls back to 1920x1080.
+ */
+export function getScreenDimensions(): { width: number; height: number } {
+  if (cachedScreenDims) return cachedScreenDims;
+  try {
+    const out = execSync(
+      `swift -e 'import CoreGraphics; let b = CGDisplayBounds(CGMainDisplayID()); print("\\(Int(b.width))x\\(Int(b.height))")'`,
+      { timeout: 10_000, encoding: 'utf-8' },
+    ).trim();
+    const [w, h] = out.split('x').map(Number);
+    if (w > 0 && h > 0) {
+      cachedScreenDims = { width: w, height: h };
+      return cachedScreenDims;
+    }
+  } catch (err) {
+    log.debug({ err }, 'Failed to query screen dimensions, using fallback');
+  }
+  return FALLBACK_SCREEN;
+}
+
+/**
+ * Find the current socket bound to a given session by reversing the
+ * `socketToSession` map. Returns `undefined` if no socket is bound.
+ */
+export function findSocketForSession(
+  sessionId: string,
+  ctx: HandlerContext,
+): net.Socket | undefined {
+  for (const [sock, id] of ctx.socketToSession) {
+    if (id === sessionId) return sock;
+  }
+  return undefined;
+}
+
+/**
+ * Wire the escalation handler on a text_qa session so that invoking
+ * `request_computer_control` creates a CU session and notifies the client.
+ *
+ * Instead of closing over the original `socket`, the handler looks up the
+ * current socket for the session at call time via `ctx.socketToSession`.
+ * This ensures the handler targets the correct socket even after a
+ * disconnect-and-rebind cycle.
+ */
+export function wireEscalationHandler(
+  session: Session,
+  _socket: net.Socket,
+  ctx: HandlerContext,
+  explicitWidth?: number,
+  explicitHeight?: number,
+): void {
+  // Import handleCuSessionCreate lazily to avoid circular dependency
+  const { handleCuSessionCreate } = require('./computer-use.js');
+
+  const dims = (explicitWidth && explicitHeight)
+    ? { width: explicitWidth, height: explicitHeight }
+    : getScreenDimensions();
+  const screenWidth = dims.width;
+  const screenHeight = dims.height;
+  session.setEscalationHandler((task: string, sourceSessionId: string): boolean => {
+    const currentSocket = findSocketForSession(sourceSessionId, ctx);
+    if (!currentSocket) {
+      log.warn({ sourceSessionId }, 'Escalation handler: no active socket found for session');
+      return false;
+    }
+
+    const cuSessionId = uuid();
+    const cuMsg: CuSessionCreate = {
+      type: 'cu_session_create',
+      sessionId: cuSessionId,
+      task,
+      screenWidth,
+      screenHeight,
+      interactionType: 'computer_use',
+    };
+    handleCuSessionCreate(cuMsg, currentSocket, ctx);
+
+    ctx.send(currentSocket, {
+      type: 'task_routed',
+      sessionId: cuSessionId,
+      interactionType: 'computer_use',
+      task,
+      escalatedFrom: sourceSessionId,
+    });
+
+    return true;
+  });
+}
+
+export function isRecord(value: unknown): value is Record<string, unknown> {
+  return typeof value === 'object' && value !== null;
+}
+
+export function formatBytes(sizeBytes: number): string {
+  if (sizeBytes < 1024) return `${sizeBytes} B`;
+  const kb = sizeBytes / 1024;
+  if (kb < 1024) return `${kb.toFixed(1)} KB`;
+  return `${(kb / 1024).toFixed(1)} MB`;
+}
+
+function clampAttachmentText(text: string): string {
+  if (text.length <= HISTORY_ATTACHMENT_TEXT_LIMIT) return text;
+  return `${text.slice(0, HISTORY_ATTACHMENT_TEXT_LIMIT)}<truncated />`;
+}
+
+function renderImageBlockForHistory(block: Record<string, unknown>): string {
+  const source = isRecord(block.source) ? block.source : null;
+  const mediaType = source && typeof source.media_type === 'string' ? source.media_type : 'image/*';
+  const sizeBytes = source && typeof source.data === 'string' ? estimateBase64Bytes(source.data) : 0;
+  if (sizeBytes <= 0) {
+    return `[Image attachment] ${mediaType}`;
+  }
+  return `[Image attachment] ${mediaType}, ${formatBytes(sizeBytes)}`;
+}
+
+function renderFileBlockForHistory(block: Record<string, unknown>): string {
+  const source = isRecord(block.source) ? block.source : null;
+  const mediaType = source && typeof source.media_type === 'string' ? source.media_type : 'application/octet-stream';
+  const filename = source && typeof source.filename === 'string' ? source.filename : 'attachment';
+  const sizeBytes = source && typeof source.data === 'string' ? estimateBase64Bytes(source.data) : 0;
+  const summaryParts = [`[File attachment] ${filename}`, `type=${mediaType}`];
+  if (sizeBytes > 0) summaryParts.push(`size=${formatBytes(sizeBytes)}`);
+
+  const extractedText = typeof block.extracted_text === 'string' ? block.extracted_text.trim() : '';
+  if (!extractedText) {
+    return summaryParts.join(', ');
+  }
+  return `${summaryParts.join(', ')}\nAttachment text: ${clampAttachmentText(extractedText)}`;
+}
+
+export function renderHistoryContent(content: unknown): RenderedHistoryContent {
+  if (!Array.isArray(content)) {
+    let text: string;
+    if (content == null) {
+      text = '';
+    } else if (typeof content === 'object') {
+      text = JSON.stringify(content);
+    } else {
+      text = String(content);
+    }
+    return {
+      text,
+      toolCalls: [],
+      toolCallsBeforeText: false,
+      textSegments: text ? [text] : [],
+      contentOrder: text ? ['text:0'] : [],
+      surfaces: [],
+    };
+  }
+
+  const textParts: string[] = [];
+  const attachmentParts: string[] = [];
+  const toolCalls: HistoryToolCall[] = [];
+  const surfaces: HistorySurface[] = [];
+  const pendingToolUses = new Map<string, HistoryToolCall>();
+  let seenText = false;
+  let seenToolUse = false;
+  let toolCallsBeforeText = false;
+
+  // Segment tracking: text blocks separated by tool_use boundaries
+  const textSegments: string[] = [];
+  const contentOrder: string[] = [];
+  let currentSegmentParts: string[] = [];
+  let hasOpenSegment = false;
+
+  function finalizeSegment(): void {
+    if (hasOpenSegment) {
+      textSegments[textSegments.length - 1] = currentSegmentParts.join('');
+      currentSegmentParts = [];
+      hasOpenSegment = false;
+    }
+  }
+
+  function ensureSegment(): void {
+    if (!hasOpenSegment) {
+      textSegments.push('');
+      contentOrder.push(`text:${textSegments.length - 1}`);
+      hasOpenSegment = true;
+    }
+  }
+
+  for (const block of content) {
+    if (!isRecord(block) || typeof block.type !== 'string') continue;
+
+    // Collect ui_surface blocks for inclusion in history
+    if (block.type === 'ui_surface') {
+      finalizeSegment();
+      const surface: HistorySurface = {
+        surfaceId: typeof block.surfaceId === 'string' ? block.surfaceId : '',
+        surfaceType: typeof block.surfaceType === 'string' ? block.surfaceType : '',
+        title: typeof block.title === 'string' ? block.title : undefined,
+        data: isRecord(block.data) ? (block.data as Record<string, unknown>) : {},
+        actions: Array.isArray(block.actions) ? block.actions : undefined,
+        display: typeof block.display === 'string' ? block.display : undefined,
+      };
+      surfaces.push(surface);
+      contentOrder.push(`surface:${surfaces.length - 1}`);
+      continue;
+    }
+
+    if (block.type === 'text' && typeof block.text === 'string') {
+      textParts.push(block.text);
+      ensureSegment();
+      currentSegmentParts.push(block.text);
+      seenText = true;
+      continue;
+    }
+    if (block.type === 'file') {
+      attachmentParts.push(renderFileBlockForHistory(block));
+      continue;
+    }
+    if (block.type === 'image') {
+      attachmentParts.push(renderImageBlockForHistory(block));
+      continue;
+    }
+    if (block.type === 'tool_use') {
+      finalizeSegment();
+      const name = typeof block.name === 'string' ? block.name : 'unknown';
+      const input = isRecord(block.input) ? block.input as Record<string, unknown> : {};
+      const id = typeof block.id === 'string' ? block.id : '';
+      const entry: HistoryToolCall = { name, input };
+      toolCalls.push(entry);
+      if (id) pendingToolUses.set(id, entry);
+      contentOrder.push(`tool:${toolCalls.length - 1}`);
+      if (!seenToolUse) {
+        seenToolUse = true;
+        if (!seenText) toolCallsBeforeText = true;
+      }
+      continue;
+    }
+    if (block.type === 'tool_result') {
+      const toolUseId = typeof block.tool_use_id === 'string' ? block.tool_use_id : '';
+      const resultContent = typeof block.content === 'string' ? block.content : '';
+      const isError = block.is_error === true;
+      // Extract base64 image data from persisted contentBlocks (e.g. browser_screenshot)
+      let imageData: string | undefined;
+      if (Array.isArray(block.contentBlocks)) {
+        const imgBlock = block.contentBlocks.find(
+          (b: Record<string, unknown>) => isRecord(b) && b.type === 'image',
+        );
+        if (imgBlock && isRecord(imgBlock) && isRecord(imgBlock.source)) {
+          const src = imgBlock.source as Record<string, unknown>;
+          if (typeof src.data === 'string') {
+            imageData = src.data;
+          }
+        }
+      }
+      const matched = toolUseId ? pendingToolUses.get(toolUseId) : null;
+      if (matched) {
+        matched.result = resultContent;
+        matched.isError = isError;
+        if (imageData) matched.imageData = imageData;
+      } else {
+        toolCalls.push({ name: 'unknown', input: {}, result: resultContent, isError, ...(imageData ? { imageData } : {}) });
+      }
+      continue;
+    }
+  }
+
+  // Attachment descriptions are NOT included in textSegments — they are
+  // rendered separately as attachment chips/images in the UI.  Including
+  // them would cause the "last text segment" heuristic to surface an
+  // attachment descriptor instead of the assistant's actual narrative text
+  // in interleaved history messages.
+
+  finalizeSegment();
+
+  const text = textParts.join('');
+  let rendered: string;
+  if (attachmentParts.length === 0) {
+    rendered = text;
+  } else if (text.trim().length === 0) {
+    rendered = attachmentParts.join('\n');
+  } else {
+    rendered = `${text}\n${attachmentParts.join('\n')}`;
+  }
+
+  return { text: rendered, toolCalls, toolCallsBeforeText, textSegments, contentOrder, surfaces };
+}
+
+export function mergeToolResults(messages: ParsedHistoryMessage[]): ParsedHistoryMessage[] {
+  // Note: We no longer merge consecutive assistant messages at load time since
+  // they are now consolidated when saved. This function only handles legacy
+  // conversations that haven't been consolidated yet, and continues to merge
+  // tool_result blocks into their preceding assistant messages.
+  const result: ParsedHistoryMessage[] = [];
+  for (let i = 0; i < messages.length; i++) {
+    const msg = messages[i];
+
+    // If this is a user message whose only content is tool_result blocks,
+    // merge those results into the preceding assistant message's toolCalls.
+    // This should rarely happen now since consolidation removes these messages,
+    // but we keep it for backwards compatibility with old data.
+    if (msg.role === 'user' && msg.text.trim() === '' && msg.toolCalls.length > 0) {
+      const prev = result.length > 0 ? result[result.length - 1] : null;
+      if (prev && prev.role === 'assistant' && prev.toolCalls.length > 0) {
+        for (const resultEntry of msg.toolCalls) {
+          const unresolved = prev.toolCalls.find(
+            (tc) => tc.result === undefined,
+          );
+          if (unresolved) {
+            unresolved.result = resultEntry.result;
+            unresolved.isError = resultEntry.isError;
+            if (resultEntry.imageData) unresolved.imageData = resultEntry.imageData;
+          }
+        }
+      }
+      // Suppress this internal user message from the visible history
+      continue;
+    }
+
+    result.push({ ...msg, toolCalls: msg.toolCalls.map((tc) => ({ ...tc })) });
+  }
+  return result;
+}
+
+/**
+ * Send a `secret_request` to the client and wait for the response,
+ * outside of a session context (e.g. from handler-level code like publish_page).
+ */
+export function requestSecretStandalone(
+  socket: net.Socket,
+  ctx: HandlerContext,
+  params: { service: string; field: string; label: string; description?: string; placeholder?: string; allowedTools?: string[] },
+): Promise<SecretPromptResult> {
+  const requestId = uuid();
+  return new Promise((resolve) => {
+    const timer = setTimeout(() => {
+      pendingStandaloneSecrets.delete(requestId);
+      resolve({ value: null, delivery: 'store' });
+    }, 120_000);
+    pendingStandaloneSecrets.set(requestId, { resolve, timer });
+    ctx.send(socket, {
+      type: 'secret_request',
+      requestId,
+      service: params.service,
+      field: params.field,
+      label: params.label,
+      description: params.description,
+      placeholder: params.placeholder,
+      purpose: 'Publish site apps to the web',
+      allowedTools: params.allowedTools,
+    });
+  });
+}
+
+/** Get or create the skill entry object for a given skill name, creating intermediate objects as needed.
+ *  Guards against malformed config (e.g. skills or entries being a string, array, or null)
+ *  by resetting non-object intermediates to {}, restoring self-healing behavior. */
+export function ensureSkillEntry(raw: Record<string, unknown>, name: string): Record<string, unknown> {
+  if (!isRecord(raw.skills) || Array.isArray(raw.skills)) raw.skills = {};
+  const skills = raw.skills as Record<string, unknown>;
+  if (!isRecord(skills.entries) || Array.isArray(skills.entries)) skills.entries = {};
+  const entries = skills.entries as Record<string, unknown>;
+  if (!isRecord(entries[name]) || Array.isArray(entries[name])) entries[name] = {};
+  return entries[name] as Record<string, unknown>;
+}
+
+/** Compare two semver strings. Returns negative if a < b, 0 if equal, positive if a > b. */
+export function compareSemver(a: string, b: string): number {
+  const pa = a.split('.').map(Number);
+  const pb = b.split('.').map(Number);
+  for (let i = 0; i < 3; i++) {
+    const diff = (pa[i] ?? 0) - (pb[i] ?? 0);
+    if (diff !== 0) return diff;
+  }
+  return 0;
+}
